% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_states.R
\name{code_modern_states}
\alias{code_modern_states}
\title{Code country specific indicators}
\usage{
code_modern_states(v)
}
\arguments{
\item{v}{A character vector leading to the COW_ID column}
}
\value{
A dataframe adding information to the pre-exported dataset.
}
\description{
Adds various country specific codes and information to the processed
dataset. This function is only useful for contemporary countries.
}
\details{
This function uses the \code{countrycode} package to identify and return
various country specific information such as the internet domain and unicode
flags to the preprocessed dataset.
}
\examples{
\dontrun{
code_modern_states(states$ISD$ID)
}
}
