% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-colrels.R
\docType{data}
\name{colrels}
\alias{colrels}
\title{colrels database documentation}
\format{
The colrels database is a list that contains the
following 1 datasets: ICOW_COL.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{ICOW_COL: }{A dataset with 221 observations and the following
8 variables: COW_ID, Label, Beg, Origin_COW_ID, IndepType,
Beg_COW, Beg_GW, Beg_Polity2.}
}
}
\usage{
colrels
}
\description{
colrels database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## $ICOW_COL
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |    COW_ID     |   numeric   |             221|     0     |       0       |
## |     Label     |  character  |             221|     0     |       0       |
## |      Beg      |    mdate    |             221|    13     |     5.88      |
## | Origin_COW_ID |   numeric   |             221|     0     |       0       |
## |   IndepType   |   numeric   |             221|     0     |       0       |
## |    Beg_COW    |    mdate    |             221|     0     |       0       |
## |    Beg_GW     |    mdate    |             221|    23     |     10.41     |
## |  Beg_Polity2  |    mdate    |             221|     0     |       0       |
## ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
