% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vdem.R
\name{extract_vdem}
\alias{extract_vdem}
\alias{import_vdem}
\alias{import_vparty}
\title{Extract V-Dem and V-Party data}
\usage{
import_vdem()

import_vparty()
}
\value{
A dataframe of the\verb{[vdem]} dataset in a many packages
universe-consistent format.

A dataframe of the\verb{[vparty]} dataset consistent with the
many packages universe.
}
\description{
Functions to import the
\href{https://www.v-dem.net/data/the-v-dem-dataset/}{V-Dem} and the
\href{https://www.v-dem.net/data/v-party-dataset/}{V-Party} datasets
from the \href{https://github.com/vdeminstitute/vdemdata}{\code{{vdemdata}}}
package in a many packages universe consistent format.
}
\details{
\code{import_vdem()} imports the latest version of VDEM dataset
and formats it to be consistent with the many packages universe.

\code{import_vparty()} imports the latest version of the V-Party dataset
and formats it to be consistent with the many packages universe.
}
