% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cshapes.R
\name{import_cshapes}
\alias{import_cshapes}
\title{Import and format \verb{[cshapes]} datasets}
\usage{
import_cshapes(date, ...)
}
\arguments{
\item{date}{The date for which the distance list should be computed.
This argument must be of type Date and must be in the range 1/1/1886 -
end of the dataset.}

\item{...}{Arguments to be passed to \verb{[cshapes]}
functions. See the \verb{[cshapes]} documentation for more details}
}
\value{
A dataframe with the qVerse-consistently formatted \verb{[cshapes]}
dataset.
}
\description{
Imports CShapes 2.0 datasets and formats them to a qVerse
consistent output.
}
\examples{
import_cshapes(date = "1900-01-01")
}
