% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cshapes.R
\name{import_distmatrix}
\alias{import_distmatrix}
\title{Imports and formats \verb{[cshapes]} pre-computed distance matrices.}
\usage{
import_distmatrix(date, type, ...)
}
\arguments{
\item{date}{The date for which the distance list should be computed.
This argument must be of type Date and must be in the range 1/1/1886 -
end of the dataset.}

\item{type}{Specifies the type of distance list: "capdist" for capital
distances, "centdist" for centroid distances, and "mindist" for minimum
distances.}

\item{...}{Arguments to be passed to \verb{[cshapes]}}
}
\value{
A matrix with the desired distance list between polygons,
capitals, or polygon centroids in kilometers.
}
\description{
\code{import_matrix()} imports a pre-computed minimum distance matrix
from the \verb{[cShapes]} package. Minimum distances are computed in three ways:
distances between capitals, distances between centroids of the polygons,
minimum distances between the polygons in kilometers.
}
\examples{
\dontrun{
import_distmatrix(date = "1900-01-01", type = "capdist")
}
}
