% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-leaders.R
\docType{data}
\name{leaders}
\alias{leaders}
\title{leaders database documentation}
\format{
The leaders database is a list that contains the
following 1 datasets: ARCHIGOS.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{ARCHIGOS: }{A dataset with 17686 observations and the following
24 variables: obsid, leadid, ccode, idacr, leader, startdate, enddate, entry, exit, exitcode, prevtimesinoffice, posttenurefate, gender, yrborn, yrdied, borndate, deathdate, dbpedia.uri, num.entry, num.exit, num.exitcode, num.posttenurefate, fties, ftcur.}
}
}
\usage{
leaders
}
\description{
leaders database documentation
}
\details{
\preformatted{## $ARCHIGOS
## -- Data Summary ------------------------
##                            Values
## Name                       X[[i]]
## Number of rows             17686 
## Number of columns          24    
## _______________________          
## Column type frequency:           
##   character                20    
##   numeric                  4     
## ________________________         
## Group variables            None  
## 
## -- Variable type: character ----------------------------------------------------------------------------------------------------------------------------------------------------------------------
## # A tibble: 20 x 8
##    skim_variable      n_missing complete_rate   min   max empty n_unique whitespace
##  * <chr>                  <int>         <dbl> <int> <int> <int>    <int>      <int>
##  1 obsid                      0         1         8    10     0     3409          0
##  2 leadid                     0         1        36    36     0     2802          0
##  3 idacr                      0         1         3     3     0      189          0
##  4 leader                     0         1         3    41     0     2757          0
##  5 startdate                  0         1        10    10     0     3279          0
##  6 enddate                    0         1        10    11     0     3155          0
##  7 entry                      0         1         7    18     0        5          0
##  8 exit                       0         1         7    25     0        9          0
##  9 exitcode                   0         1         1    59     0       16          0
## 10 prevtimesinoffice          0         1         1     1     0        8          0
## 11 posttenurefate             0         1         1    57     0       10          0
## 12 gender                     0         1         1     1     0        2          0
## 13 borndate                5179         0.707    10    12     0     1656          0
## 14 deathdate              11762         0.335    10    10     0      659          0
## 15 dbpedia.uri             8552         0.516    32    72     0     1158          0
## 16 num.entry                  0         1         1     4     0        5          0
## 17 num.exit                   0         1         1     4     0        8          0
## 18 num.exitcode               0         1         1     4     0       16          0
## 19 num.posttenurefate         0         1         1     4     0       10          0
## 20 fties                  15814         0.106    48   190     0      200          0
## 
## -- Variable type: numeric ------------------------------------------------------------------------------------------------------------------------------------------------------------------------
## # A tibble: 4 x 10
##   skim_variable n_missing complete_rate     mean       sd    p0   p25   p50   p75  p100
## * <chr>             <int>         <dbl>    <dbl>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
## 1 ccode                 0        1       413.     254.        2   200   375   645   950
## 2 yrborn                0        1      1894.     172.     -999  1874  1912  1937  1983
## 3 yrdied                0        1      1040.    1291.     -999  -777  1925  1975  2015
## 4 ftcur             16031        0.0936    0.389    0.488     0     0     0     1     1
}
}
\keyword{datasets}
