% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-package.R
\docType{package}
\name{manystates-package}
\alias{manystates}
\alias{manystates-package}
\title{manystates: Many Data on State and State-Like Actors in the International System}
\description{
Comprehensively identifying states and state-like actors is difficult. This package provides data on states and state-like entities in the international system across time. The package combines and cross-references several existing datasets consistent with the aims and functions of the manydata package. It also includes functions for identifying state references in text, and for generating fictional state names.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://globalgov.github.io/manystates/}
  \item Report bugs at \url{https://github.com/globalgov/manystates/issues}
}

}
\author{
\strong{Maintainer}: James Hollway \email{james.hollway@graduateinstitute.ch} (\href{https://orcid.org/0000-0002-8361-9647}{ORCID}) (IHEID) [contributor]

Other contributors:
\itemize{
  \item Bernhard Bieri (\href{https://orcid.org/0000-0001-5943-9059}{ORCID}) (IHEID) [contributor]
  \item Mylan Evrard (\href{https://orcid.org/0009-0009-4404-1078}{ORCID}) (IHEID) [contributor]
  \item Esther Peev (\href{https://orcid.org/0000-0002-9678-2777}{ORCID}) (IHEID) [contributor]
  \item Henrique Sposito (\href{https://orcid.org/0000-0003-3420-6085}{ORCID}) (IHEID) [contributor]
  \item Jael Tan (\href{https://orcid.org/0000-0002-6234-9764}{ORCID}) (IHEID) [contributor]
}

}
\keyword{internal}
