% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_names.R
\name{namesta}
\alias{namesta}
\title{Stardadizes State Names (Function partially taken from Gnevar Package)}
\usage{
namesta(id, strict = FALSE)
}
\arguments{
\item{id}{A string}

\item{strict}{By default FALSE}
}
\value{
A standard state name across datasets
}
\description{
Stardadizes State Names (Function partially taken from Gnevar Package)
}
\details{
The function standardize state names as well as capitalises all words in the strings passed to it,
trimms all white space from the starts and ends of the strings, remove points and signs.
}
