% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qStates-package.R
\docType{package}
\name{qStates-package}
\alias{qStates}
\alias{qStates-package}
\title{qStates: States for qData}
\description{
qStates provide data on states in the international system across time. 
    The package combines and cross-references several existing datasets consistent with the qData package aims.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/globalgov/qStates}
  \item Report bugs at \url{https://github.com/globalgov/qStates/issues}
}

}
\author{
\strong{Maintainer}: James Hollway \email{james.hollway@graduateinstitute.ch} (\href{https://orcid.org/0000-0002-8361-9647}{ORCID}) (IHEID) [contributor]

Other contributors:
\itemize{
  \item Henrique Sposito (IHEID) [contributor]
}

}
\keyword{internal}
