% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-ratrules.R
\docType{data}
\name{ratrules}
\alias{ratrules}
\title{ratrules database documentation}
\format{
The ratrules database is a list that contains the
following 1 datasets: RATRULES.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{RATRULES: }{A dataset with 215 observations and the following
5 variables: stateID, Label, RatProcedure, ConstitutionalDesc, url.}
}
}
\usage{
ratrules
}
\description{
ratrules database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## $RATRULES
## -------------------------------------------------------------------------------
## |   Column Name    |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## -------------------------------------------------------------------------------
## |     stateID      |  character  |             215|     0     |       0       |
## |      Label       |  character  |             215|     0     |       0       |
## |   RatProcedure   |   numeric   |             215|    15     |     6.98      |
## |ConstitutionalDesc|  character  |             215|    192    |     89.3      |
## |       url        |  character  |             215|    185    |     86.05     |
## -------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
