% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_system.R
\name{state_system}
\alias{state_system}
\title{Obtaining a list of state-like entities from a datacube at a certain point in time}
\usage{
state_system(datacube, date = Sys.Date())
}
\arguments{
\item{datacube}{A datacube,
i.e. a list of data frames with Begin and End date variables.}

\item{date}{A date (of class Date or character) at which to filter the datacube.}
}
\description{
Obtaining a list of state-like entities from a datacube at a certain point in time
}
\examples{
state_system(manystates::states, date = "2010")
}
