% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_states.R
\name{code_states}
\alias{code_states}
\title{Code Agreement Parties}
\usage{
code_states(title, activity = TRUE, replace = NULL)
}
\arguments{
\item{title}{A character vector of treaty titles}

\item{activity}{Do you want the activity of treaty to be coded?
By default, TRUE.}

\item{replace}{Do you want the state name or abbreviation to be returned?
By default, NULL.
Other options include, "names", for the state name,
or "ID", for the 3 letter state abbreviation.}
}
\value{
A character vector of parties
that are mentioned in the treaty title
}
\description{
Identify the countries that are part of an agreement.
}
\details{
The function codes states in treaty titles.
The function also returns the "activity" for bilateral treaties coded,
if activity is TRUE.
Bilateral agreements usually detail their activity and specify area in the
last words of the titles.
These last words are abbreviated by the function to differentiate between
bilateral treaties and avoid false positives being generated since multiple,
different, bilateral treaties are often signed in the same day.
For the complete list of parties coded please run the function without
an argument (i.e. \code{code_states()}).
}
\examples{
\dontrun{
IEADB <- dplyr::slice_sample(manyenviron::agreements$IEADB, n = 10)
code_states(IEADB$Title)
code_states(IEADB$Title, activity = FALSE)
code_states(IEADB$Title, activity = FALSE, replace = "names")
code_states(IEADB$Title, activity = FALSE, replace = "ID")
}
}
