% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_states.R
\name{code_states}
\alias{code_states}
\title{Code parties}
\usage{
code_states(v)
}
\arguments{
\item{v}{A character vector}
}
\value{
A character vector of parties, in English, separated by commas.
}
\description{
Retrieves countries from a character vector
}
\details{
This function builds upon the \code{stat_actor} list and the
\code{countrycode} package to identify and return the parties mentioned
in a character vector of agreement titles or texts.
}
\examples{
states <- c("Two are from Switzerland", "One from New Zealand", "And one from Brazil")
code_states(states)
}
