% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-contiguity.R
\docType{data}
\name{contiguity}
\alias{contiguity}
\title{State contiguity datacube}
\format{
\describe{
\item{COW: }{A dataset with 847
observations and 12 variables:
stateID1, stateID2, Begin, End, StateName1, StateName2, ContiguityType, dyadID, cowID1, cowNR1, cowID2, and cowNR2.}
\item{HUGGO: }{A dataset with 609
observations and 8 variables:
stateID1, stateID2, Begin, End, StateName1, StateName2, ContiguityType, and url.}
}
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
}
\source{
\itemize{
\item{COW: }{
Hensel, Paul R. \emph{Correlates of War Project. Direct Contiguity Data, 1816-2016. Version 3.2.} 2017.
\url{https://correlatesofwar.org/data-sets/direct-contiguity}
}
\item{HUGGO: }{
J. Hollway. manystates: States for manydata. 2021.}
}
}
\usage{
contiguity
}
\description{
The contiguity datacube is a list containing
2 datasets:
HUGGO and COW.
It is a work-in-progress, so please do not rely on it yet.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $HUGGO
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |   stateID1    |  character  |             609|     0     |       0       |
#> |   stateID2    |  character  |             609|     3     |     0.49      |
#> |     Begin     |    mdate    |             609|     0     |       0       |
#> |      End      |    mdate    |             609|     0     |       0       |
#> |  StateName1   |  character  |             609|     0     |       0       |
#> |  StateName2   |  character  |             609|     0     |       0       |
#> |ContiguityType |   numeric   |             609|     0     |       0       |
#> |      url      |  character  |             609|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $COW
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |   stateID1    |  character  |             847|     0     |       0       |
#> |   stateID2    |  character  |             847|     0     |       0       |
#> |     Begin     |    mdate    |             847|     0     |       0       |
#> |      End      |    mdate    |             847|     0     |       0       |
#> |  StateName1   |  character  |             847|     0     |       0       |
#> |  StateName2   |  character  |             847|     0     |       0       |
#> |ContiguityType |   numeric   |             847|     0     |       0       |
#> |    dyadID     |   numeric   |             847|     0     |       0       |
#> |    cowID1     |  character  |             847|     0     |       0       |
#> |    cowNR1     |   numeric   |             847|     0     |       0       |
#> |    cowID2     |  character  |             847|     0     |       0       |
#> |    cowNR2     |   numeric   |             847|     0     |       0       |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{Mapping}{

\itemize{
\item{COW_CONT: }{
Variable Mapping\tabular{ll}{
   \emph{manystates} \tab \emph{COW} \cr
   stateID1 \tab statelab \cr
   stateID2 \tab statehab \cr
   Begin \tab begin \cr
   End \tab end \cr
   dyadID \tab dyad \cr
   cowNR1 \tab statelno \cr
   cowNR2 \tab statehno \cr
   ContiguityType \tab conttype \cr
}


}
}
}

\keyword{datasets}
