% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-contiguity.R
\docType{data}
\name{contiguity}
\alias{contiguity}
\title{contiguity database documentation}
\format{
The contiguity database is a list that contains the
following 2 datasets: DIRCONT, GNEVAR_CONT.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{DIRCONT: }{A dataset with 847 observations and the following
8 variables: dyadID, ContiguityType, Beg, End, stateID1, Label1, stateID2, Label2.}
\item{GNEVAR_CONT: }{A dataset with 192 observations and the following
9 variables: stateID, Label, Beg, End, Contiguity, EntityType, FAOmembership,
Group, url.}
}
}
\usage{
contiguity
}
\description{
contiguity database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $DIRCONT
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    dyadID     |   numeric   |             847|     0     |       0       |
#> |ContiguityType |   numeric   |             847|     0     |       0       |
#> |      Beg      |    mdate    |             847|     0     |       0       |
#> |      End      |    mdate    |             847|     0     |       0       |
#> |   stateID1    |   numeric   |             847|     0     |       0       |
#> |    Label1     |  character  |             847|     0     |       0       |
#> |   stateID2    |   numeric   |             847|     0     |       0       |
#> |    Label2     |  character  |             847|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $GNEVAR_CONT
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    stateID    |  character  |             192|     0     |       0       |
#> |     Label     |  character  |             192|     0     |       0       |
#> |      Beg      |    mdate    |             192|     0     |       0       |
#> |      End      |    mdate    |             192|     0     |       0       |
#> |  Contiguity   |  character  |             192|    41     |     21.35     |
#> |  EntityType   |  character  |             192|     0     |       0       |
#> | FAOmembership |  character  |             192|     0     |       0       |
#> |     Group     |  character  |             192|     0     |       0       |
#> |      url      |  character  |             192|     0     |       0       |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
