% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cshapes.R
\name{extract_cshapes}
\alias{extract_cshapes}
\alias{import_cshapes}
\alias{import_distlist}
\alias{import_distmatrix}
\title{Extract CShapes data and matrix}
\usage{
import_cshapes(date, ...)

import_distlist(date, type, ...)

import_distmatrix(date, type, ...)
}
\arguments{
\item{date}{The date for which the distance list should be computed.
This argument must be of type Date and must be in the range 1/1/1886 -
end of the dataset.}

\item{...}{Arguments to be passed to \verb{[cshapes]} functions.
See the \verb{[cshapes]} documentation for more details}

\item{type}{Specifies the type of distance list: "capdist" for capital
distances, "centdist" for centroid distances, and "mindist" for minimum
distances.}
}
\value{
A dataframe with the qVerse-consistently formatted \verb{[cshapes]}
dataset.

A dataframe with the desired distance list between polygons,
capitals, or polygon centroids in kilometers.

A matrix with the desired distance list between polygons,
capitals, or polygon centroids in kilometers.
}
\description{
Functions to import CShapes 2.0 datasets, distances and
formats them to a qVerse consistent output from the \verb{[cShapes]} package.
}
\details{
\code{import_cshapes()}imports CShapes 2.0 datasets
and formats them to a qVerse consistent output.

\code{import_distlist()} imports pre-computed minimum
distance dataframe from the \verb{[cShapes]} package.
Minimum distances are computed in three ways:
distances between capitals,
distances between centroids of the polygons,
minimum distances between the polygons in kilometers.

\code{import_distlist()} imports and formats pre-computed
minimum distance matrix from the \verb{[cShapes]} package.
Minimum distances are computed in three ways:
distances between capitals,
distances between centroids of the polygons,
minimum distances between the polygons in kilometers.
}
\examples{
\donttest{
import_cshapes(date = "1900-01-01")
}
\donttest{
import_distlist(date = "1900-01-01", type = "capdist")
}
\donttest{
import_distmatrix(date = "1900-01-01", type = "capdist")
}
}
