% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vdem.R
\name{extract_vdem}
\alias{extract_vdem}
\alias{import_vdem}
\alias{import_vparty}
\title{Extract V-Dem and V-Party data}
\usage{
import_vdem()

import_vparty()
}
\value{
A dataframe of the\verb{[vdem]} dataset in a many packages
universe-consistent format.

A dataframe of the\verb{[vparty]} dataset in a many packages universe
consistent format.
}
\description{
Functions to import the
\href{https://www.v-dem.net/en/data/data/v-dem-dataset-v111/}{V-Dem} and the
\href{https://www.v-dem.net/en/data/data/v-party-dataset/}{V-Party} datasets
from their \href{https://github.com/vdeminstitute/vdemdata}{\code{{vdemdata}}}
package in a many packages universe consistent format.
}
\details{
\code{import_vdem()} imports VDem 11.1 dataset
and formats them to a many packages universe consistent output.

\code{import_vparty()} imports the V-Party dataset and formats it to
a many packages universe consistent dataframe.
}
\examples{
\donttest{
import_vdem()
}
\donttest{
import_vparty()
}
}
