% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_states.R
\name{syllabise_states}
\alias{syllabise_states}
\alias{syllabize_states}
\alias{generate_states}
\title{Generate a list of fictional country names}
\usage{
generate_states(n = 10, prefixed = 0.15, suffixed = 0.15)
}
\arguments{
\item{word}{One or more words (character vector) to split into syllable-like units.}

\item{prefixed}{Proportion of country names with a prefix by default 0.15.}

\item{suffixed}{Proportion of country names with a suffix by default 0.15.}
}
\value{
String vector of fictional country names
}
\description{
Generate a list of fictional country names
}
\examples{
  generate_states(12)
}
