% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_states.R
\name{generate_states}
\alias{generate_states}
\title{Generate a list of fictional country names}
\usage{
generate_states(n = 10, prefixed = 0.15, suffixed = 0.15)
}
\arguments{
\item{n}{Integer number of country names to generate
from a library of fictional country names.}

\item{affixed}{Proportion of country names with either
a prefix or a suffix, by default 0.3.}
}
\value{
String vector of fictional country names
}
\description{
Generate a list of fictional country names
}
\examples{
  generate_states(12)
}
