% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-genevar.R
\docType{data}
\name{genevar}
\alias{genevar}
\title{genevar database documentation}
\format{
The genevar database is a list that contains the
following 1 datasets: ARCHIGOSgenevar.
This database is a little special since it contains the hand coded
alterations made by the qVerse team of the different datasets in the package.
For more information and references to each of the datasets used,
please use the \code{data_source()}, \code{data_contrast()} functions.
\describe{
\item{ARCHIGOSgenevar: }{A dataset with 24 observations and the following
30 variables: ID, LeadID, ccode, idacr, Label, leader, Beg, End, BornDate,
DeathDate, YearBorn, YearDied, Female, entry, exit, exitcode,
prevtimesinoffice, posttenurefate, dbpedia.uri, num.entry, num.exit,
num.exitcode, num.posttenurefate, FtiesNameA, FtiesCodeA, FtiesNameB,
FtiesCodeB, FtiesNameC, FtiesCodeC, ftcur.}
}
}
\usage{
genevar
}
\description{
genevar database documentation
}
\keyword{datasets}
