% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-leaders.R
\docType{data}
\name{leaders}
\alias{leaders}
\title{leaders database documentation}
\format{
The leaders database is a list that contains the
following 1 datasets: ARCHIGOS.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{ARCHIGOS: }{A dataset with 17686 observations and the following
30 variables: ARCHIGOS_ID, LeadID, COW_ID, idacr, Label, leader, Beg, End,
BornDate, DeathDate, YearBorn, YearDied, Female, entry, exit, exitcode,
prevtimesinoffice, posttenurefate, dbpedia.uri, num.entry, num.exit,
num.exitcode, num.posttenurefate, FtiesNameA, FtiesCodeA, FtiesNameB,
FtiesCodeB, FtiesNameC, FtiesCodeC, ftcur.}
}
}
\usage{
leaders
}
\description{
leaders database documentation
}
\details{
\preformatted{## $ARCHIGOS
## -- Data Summary ------------------------
##                            Values
## Name                       X[[i]]
## Number of rows             17686 
## Number of columns          30    
## _______________________          
## Column type frequency:           
##   character                28    
##   numeric                  2     
## ________________________         
## Group variables            None  
## 
## -- Variable type: character ----------------------------------------------------
## # A tibble: 28 x 8
##    skim_variable      n_missing complete_rate   min   max empty n_unique
##  * <chr>                  <int>         <dbl> <int> <int> <int>    <int>
##  1 ARCHIGOS_ID                0      1            8    10     0     3409
##  2 LeadID                     0      1           36    36     0     2802
##  3 COW_ID                     0      1            3     3     0      188
##  4 idacr                      0      1            3     3     0      189
##  5 Label                      0      1            4    32     0      187
##  6 leader                     0      1            3    41     0     2757
##  7 Beg                        0      1           10    10     0     3279
##  8 End                        0      1           10    10     0     3155
##  9 BornDate                5179      0.707       10    12     0     1656
## 10 DeathDate              11762      0.335       10    10     0      659
## 11 YearBorn                   0      1            4     5     0      175
## 12 YearDied                   0      1            4     5     0      143
## 13 entry                  14277      0.193        7    18     0        4
## 14 exit                   14277      0.193        7    25     0        8
## 15 exitcode               14277      0.193        7    59     0       15
## 16 prevtimesinoffice      14277      0.193        1     1     0        7
## 17 posttenurefate         14324      0.190        2    57     0        8
## 18 dbpedia.uri             8552      0.516       32    72     0     1158
## 19 num.entry              14277      0.193        1     4     0        4
## 20 num.exit               16902      0.0443       1     4     0        7
## 21 num.exitcode           14277      0.193        1     4     0       15
## 22 num.posttenurefate     14277      0.193        1     4     0        9
## 23 FtiesNameA             15814      0.106       11    43     0      200
## 24 FtiesCodeA             15814      0.106       36    36     0      195
## 25 FtiesNameB             17637      0.00277     15    38     0        7
## 26 FtiesCodeB             17637      0.00277     36    36     0        6
## 27 FtiesNameC             17674      0.000679    29    29     0        1
## 28 FtiesCodeC             17674      0.000679    36    36     0        1
##    whitespace
##  *      <int>
##  1          0
##  2          0
##  3          0
##  4          0
##  5          0
##  6          0
##  7          0
##  8          0
##  9          0
## 10          0
## 11          0
## 12          0
## 13          0
## 14          0
## 15          0
## 16          0
## 17          0
## 18          0
## 19          0
## 20          0
## 21          0
## 22          0
## 23          0
## 24          0
## 25          0
## 26          0
## 27          0
## 28          0
## 
## -- Variable type: numeric ------------------------------------------------------
## # A tibble: 2 x 10
##   skim_variable n_missing complete_rate   mean    sd    p0   p25   p50   p75
## * <chr>             <int>         <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
## 1 Female                0        1      0.0183 0.134     0     0     0     0
## 2 ftcur             16031        0.0936 0.389  0.488     0     0     0     1
##    p100
## * <dbl>
## 1     1
## 2     1
}
}
\keyword{datasets}
