% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-leaders.R
\docType{data}
\name{leaders}
\alias{leaders}
\title{leaders database documentation}
\format{
The leaders database is a list that contains the
following 2 datasets: ARCHIGOS, GNEVAR_ARCHIGOS.
GNEVAR_* datasets are handcoded original data or
corrections to existing datasets where coding errors were found.
In the latter case, the suffix will reference the original dataset.
In the leaders database, GNEVAR_ARCHIGOS contains corrections to
birth dates and familial ties of states' leaders.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{ARCHIGOS: }{A dataset with 17686 observations and the following
30 variables: archigosID, leaderID, cowID, idacr, Label, leader, Beg, End,
BornDate, DeathDate, YearBorn, YearDied, Female, entry, exit, exitcode,
prevtimesinoffice, posttenurefate, dbpedia.uri, num.entry, num.exit,
num.exitcode, num.posttenurefate, FtiesNameA, FtiesCodeA, FtiesNameB,
FtiesCodeB, FtiesNameC, FtiesCodeC, ftcur.}
\item{GNEVAR_ARCHIGOS: }{A dataset with 24 observations and the following
30 variables: archigosID, cowID, leaderID, idacr, Label, leader, Beg, End,
BornDate, DeathDate, YearBorn, YearDied, Female, entry, exit, exitcode,
prevtimesinoffice, posttenurefate, dbpedia.uri, num.entry, num.exit,
num.exitcode, num.posttenurefate, FtiesNameA, FtiesCodeA, FtiesNameB,
FtiesCodeB, FtiesNameC, FtiesCodeC, ftcur.}
}
}
\usage{
leaders
}
\description{
leaders database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $ARCHIGOS
#> -------------------------------------------------------------------------------
#> |   Column Name    |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> -------------------------------------------------------------------------------
#> |    archigosID    |  character  |           17686|     0     |       0       |
#> |     leaderID     |  character  |           17686|     0     |       0       |
#> |      cowID       |  character  |           17686|     0     |       0       |
#> |      idacr       |  character  |           17686|     0     |       0       |
#> |      Label       |  character  |           17686|     0     |       0       |
#> |      leader      |  character  |           17686|     0     |       0       |
#> |       Beg        |    mdate    |           17686|     0     |       0       |
#> |       End        |    mdate    |           17686|     0     |       0       |
#> |     BornDate     |    mdate    |           17686|   5179    |     29.28     |
#> |    DeathDate     |    mdate    |           17686|   11762   |     66.5      |
#> |     YearBorn     |    mdate    |           17686|     0     |       0       |
#> |     YearDied     |    mdate    |           17686|     0     |       0       |
#> |      Female      |   numeric   |           17686|     0     |       0       |
#> |      entry       |  character  |           17686|   14277   |     80.72     |
#> |       exit       |  character  |           17686|   14277   |     80.72     |
#> |     exitcode     |  character  |           17686|   14277   |     80.72     |
#> |prevtimesinoffice |  character  |           17686|   14277   |     80.72     |
#> |  posttenurefate  |  character  |           17686|   14324   |     80.99     |
#> |   dbpedia.uri    |  character  |           17686|   8552    |     48.35     |
#> |    num.entry     |  character  |           17686|   14277   |     80.72     |
#> |     num.exit     |  character  |           17686|   16902   |     95.57     |
#> |   num.exitcode   |  character  |           17686|   14277   |     80.72     |
#> |num.posttenurefate|  character  |           17686|   14277   |     80.72     |
#> |    FtiesNameA    |  character  |           17686|   15814   |     89.42     |
#> |    FtiesCodeA    |  character  |           17686|   15814   |     89.42     |
#> |    FtiesNameB    |  character  |           17686|   17637   |     99.72     |
#> |    FtiesCodeB    |  character  |           17686|   17637   |     99.72     |
#> |    FtiesNameC    |  character  |           17686|   17674   |     99.93     |
#> |    FtiesCodeC    |  character  |           17686|   17674   |     99.93     |
#> |      ftcur       |   integer   |           17686|   16031   |     90.64     |
#> -------------------------------------------------------------------------------
#> 
#> 
#> $GNEVAR_ARCHIGOS
#> -------------------------------------------------------------------------------
#> |   Column Name    |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> -------------------------------------------------------------------------------
#> |    archigosID    |  character  |              24|     0     |       0       |
#> |      cowID       |  character  |              24|     0     |       0       |
#> |     leaderID     |  character  |              24|     0     |       0       |
#> |      idacr       |  character  |              24|     0     |       0       |
#> |      Label       |  character  |              24|     0     |       0       |
#> |      leader      |  character  |              24|     0     |       0       |
#> |       Beg        |    mdate    |              24|     0     |       0       |
#> |       End        |    mdate    |              24|     0     |       0       |
#> |     BornDate     |    mdate    |              24|    10     |     41.67     |
#> |    DeathDate     |    mdate    |              24|    11     |     45.83     |
#> |     YearBorn     |    mdate    |              24|     0     |       0       |
#> |     YearDied     |    mdate    |              24|     0     |       0       |
#> |      Female      |   numeric   |              24|     0     |       0       |
#> |      entry       |  character  |              24|    18     |      75       |
#> |       exit       |  character  |              24|    18     |      75       |
#> |     exitcode     |  character  |              24|    18     |      75       |
#> |prevtimesinoffice |  character  |              24|    18     |      75       |
#> |  posttenurefate  |  character  |              24|    18     |      75       |
#> |   dbpedia.uri    |  character  |              24|    11     |     45.83     |
#> |    num.entry     |  character  |              24|    18     |      75       |
#> |     num.exit     |  character  |              24|    23     |     95.83     |
#> |   num.exitcode   |  character  |              24|    18     |      75       |
#> |num.posttenurefate|  character  |              24|    18     |      75       |
#> |    FtiesNameA    |  character  |              24|     6     |      25       |
#> |    FtiesCodeA    |  character  |              24|     6     |      25       |
#> |    FtiesNameB    |  character  |              24|    24     |      100      |
#> |    FtiesCodeB    |  character  |              24|    19     |     79.17     |
#> |    FtiesNameC    |  character  |              24|    24     |      100      |
#> |    FtiesCodeC    |  character  |              24|    24     |      100      |
#> |      ftcur       |   integer   |              24|     6     |      25       |
#> -------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
