% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qStates-leaders.R
\docType{data}
\name{leaders}
\alias{leaders}
\title{Leaders database documentation}
\format{
The leaders database is a list that contains the
following 1 datasets: ARCHIGOS.
For more information and references to each of the datasets used,
please use the \code{data_source()}, \code{data_contrast()}, and \code{data_evolution()}
functions.
\describe{
\item{ARCHIGOS: }{A dataset with 17686 observations and the following
30 variables: ID, LeadID, ccode, idacr, Label, leader, Beg, End, BornDate, DeathDate, YearBorn, YearDied, Female, entry, exit, exitcode, prevtimesinoffice, posttenurefate, dbpedia.uri, num.entry, num.exit, num.exitcode, num.posttenurefate, FtiesNameA, FtiesCodeA, FtiesNameB, FtiesCodeB, FtiesNameC, FtiesCodeC, ftcur.}
}
}
\usage{
leaders
}
\description{
Leaders database documentation
}
\keyword{datasets}
