% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-regimes.R
\docType{data}
\name{regimes}
\alias{regimes}
\title{regimes database documentation}
\format{
The regimes database is a list that contains the
following 2 datasets: FreedomHouse, Polity5.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{FreedomHouse: }{A dataset with 10680 observations and the following
47 variables: ID, cowID, Year, Label, Region, Edition, Status, PR rating,
CL rating, A1, A2, A3, A, B1, B2, B3, B4, B, C1, C2, C3, C, Add Q, Add A,
PR, D1, D2, D3, D4, D, E1, E2, E3, E, F1, F2, F3, F4, F, G1, G2, G3, G4,
G, CL, Total, Territory.}
\item{Polity5: }{A dataset with 17574 observations and the following
35 variables: ID, cowID, Year, Label, p5, cyear, ccode, flag, fragment,
democ, autoc, polity, polity2, durable, xrreg, xrcomp, xropen, xconst,
parreg, parcomp, exrec, exconst, polcomp, prior, eprec, interim, bprec, post,
change, d5, sf, regtrans, Beg, End, speccat.}
}
}
\usage{
regimes
}
\description{
regimes database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{## $FreedomHouse
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |      ID       |  character  |           10680|     0     |       0       |
## |     cowID     |  character  |           10680|    377    |     3.53      |
## |     Year      |    mdate    |           10680|     0     |       0       |
## |     Label     |  character  |           10680|     0     |       0       |
## |    Region     |  character  |           10680|   7320    |     68.54     |
## |    Edition    |    mdate    |           10680|   7320    |     68.54     |
## |    Status     |  character  |           10680|   2157    |     20.2      |
## |   PR rating   |   numeric   |           10680|   1539    |     14.41     |
## |   CL rating   |   numeric   |           10680|   1539    |     14.41     |
## |      A1       |   numeric   |           10680|   8795    |     82.35     |
## |      A2       |   numeric   |           10680|   8795    |     82.35     |
## |      A3       |   numeric   |           10680|   8795    |     82.35     |
## |       A       |   numeric   |           10680|   7320    |     68.54     |
## |      B1       |   numeric   |           10680|   8795    |     82.35     |
## |      B2       |   numeric   |           10680|   8795    |     82.35     |
## |      B3       |   numeric   |           10680|   8795    |     82.35     |
## |      B4       |   numeric   |           10680|   8795    |     82.35     |
## |       B       |   numeric   |           10680|   7320    |     68.54     |
## |      C1       |   numeric   |           10680|   8795    |     82.35     |
## |      C2       |   numeric   |           10680|   8795    |     82.35     |
## |      C3       |   numeric   |           10680|   8795    |     82.35     |
## |       C       |   numeric   |           10680|   7320    |     68.54     |
## |     Add Q     |   numeric   |           10680|   7320    |     68.54     |
## |     Add A     |   numeric   |           10680|   8164    |     76.44     |
## |      PR       |   numeric   |           10680|   7320    |     68.54     |
## |      D1       |   numeric   |           10680|   8795    |     82.35     |
## |      D2       |   numeric   |           10680|   8795    |     82.35     |
## |      D3       |   numeric   |           10680|   8795    |     82.35     |
## |      D4       |   numeric   |           10680|   8795    |     82.35     |
## |       D       |   numeric   |           10680|   7320    |     68.54     |
## |      E1       |   numeric   |           10680|   8795    |     82.35     |
## |      E2       |   numeric   |           10680|   8795    |     82.35     |
## |      E3       |   numeric   |           10680|   8795    |     82.35     |
## |       E       |   numeric   |           10680|   7320    |     68.54     |
## |      F1       |   numeric   |           10680|   8795    |     82.35     |
## |      F2       |   numeric   |           10680|   8795    |     82.35     |
## |      F3       |   numeric   |           10680|   8795    |     82.35     |
## |      F4       |   numeric   |           10680|   8795    |     82.35     |
## |       F       |   numeric   |           10680|   7320    |     68.54     |
## |      G1       |   numeric   |           10680|   8795    |     82.35     |
## |      G2       |   numeric   |           10680|   8795    |     82.35     |
## |      G3       |   numeric   |           10680|   8795    |     82.35     |
## |      G4       |   numeric   |           10680|   8795    |     82.35     |
## |       G       |   numeric   |           10680|   7320    |     68.54     |
## |      CL       |   numeric   |           10680|   7320    |     68.54     |
## |     Total     |   numeric   |           10680|   6702    |     62.75     |
## |   Territory   |   numeric   |           10680|     0     |       0       |
## ----------------------------------------------------------------------------
## 
## 
## $Polity5
## ----------------------------------------------------------------------------
## |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
## ----------------------------------------------------------------------------
## |      ID       |  character  |           17574|     0     |       0       |
## |     cowID     |  character  |           17574|     0     |       0       |
## |     Year      |    mdate    |           17574|     0     |       0       |
## |     Label     |  character  |           17574|     0     |       0       |
## |      p5       |   numeric   |           17574|     0     |       0       |
## |     cyear     |   numeric   |           17574|     0     |       0       |
## |     ccode     |   numeric   |           17574|     0     |       0       |
## |     flag      |   numeric   |           17574|     0     |       0       |
## |   fragment    |   numeric   |           17574|   14128   |     80.39     |
## |     democ     |   numeric   |           17574|    809    |      4.6      |
## |     autoc     |   numeric   |           17574|    809    |      4.6      |
## |    polity     |   numeric   |           17574|    809    |      4.6      |
## |    polity2    |   numeric   |           17574|    268    |     1.52      |
## |    durable    |   numeric   |           17574|   1264    |     7.19      |
## |     xrreg     |   numeric   |           17574|    809    |      4.6      |
## |    xrcomp     |   numeric   |           17574|    809    |      4.6      |
## |    xropen     |   numeric   |           17574|    809    |      4.6      |
## |    xconst     |   numeric   |           17574|    809    |      4.6      |
## |    parreg     |   numeric   |           17574|    809    |      4.6      |
## |    parcomp    |   numeric   |           17574|    809    |      4.6      |
## |     exrec     |   numeric   |           17574|    183    |     1.04      |
## |    exconst    |   numeric   |           17574|    809    |      4.6      |
## |    polcomp    |   numeric   |           17574|    126    |     0.72      |
## |     prior     |   numeric   |           17574|   16195   |     92.15     |
## |     eprec     |   numeric   |           17574|   16121   |     91.73     |
## |    interim    |   numeric   |           17574|   16284   |     92.66     |
## |     bprec     |   numeric   |           17574|   16287   |     92.68     |
## |     post      |   numeric   |           17574|   16302   |     92.76     |
## |    change     |   numeric   |           17574|   15346   |     87.32     |
## |      d5       |   numeric   |           17574|   16277   |     92.62     |
## |      sf       |   numeric   |           17574|   17276   |     98.3      |
## |   regtrans    |   numeric   |           17574|   15334   |     87.25     |
## |      Beg      |    mdate    |           17574|     0     |       0       |
## |      End      |    mdate    |           17574|     0     |       0       |
## |    speccat    |   numeric   |           17574|   16765   |     95.4      |
## ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
