% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qStates-regimes.R
\docType{data}
\name{regimes}
\alias{regimes}
\title{Regimes database documentation}
\format{
The regimes database is a list that contains the
following 1 datasets: Polity5.
For more information and references to each of the datasets used,
please use the \code{data_source()}, \code{data_contrast()}, and \code{data_evolution()}
functions.
\describe{
\item{Polity5: }{A dataset with 17574 observations and the following
33 variables: ID, year, Label, Beg, End, p5, cyear, flag, fragment, democ, autoc, polity, polity2, durable, xrreg, xrcomp, xropen, xconst, parreg, parcomp, exrec, exconst, polcomp, prior, eprec, interim, bprec, post, change, d5, sf, regtrans, speccat.}
}
}
\usage{
regimes
}
\description{
Regimes database documentation
}
\keyword{datasets}
