% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qStates-regimes.R
\docType{data}
\name{regimes}
\alias{regimes}
\title{regimes database documentation}
\format{
The regimes database is a list that contains the
following 2 datasets: Polity5, Polity5d.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()}.
\describe{
\item{Polity5: }{A dataset with 17574 observations and the
following 32 variables: ID, Beg, End, Label, p5,
cyear, year, flag, fragment, democ, autoc, polity,
polity2, durable, xrreg, xrcomp, xropen, xconst,
parreg, parcomp, exrec, exconst, polcomp, prior,
eprec, interim, bprec, post, change, d5, sf, regtrans.}
\item{Polity5d: }{A dataset with 1692 observations and the
following 19 variables: ID, Beg, End, Label, present,
p5, polity, persist, democ, autoc, xrreg, xrcomp,
xropen, xconst, parreg, parcomp, exrec, exconst, polcomp.}
}
}
\usage{
regimes
}
\description{
regimes database documentation
}
\keyword{datasets}
