% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-regimes.R
\docType{data}
\name{regimes}
\alias{regimes}
\title{regimes database documentation}
\format{
The regimes database is a list that contains the
following 4 datasets: Polity5, FreedomHouse1, FreedomHouse2, FreedomHouse3.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{Polity5: }{A dataset with 17574 observations and the following
35 variables: ID, COW_ID, Year, Label, p5, cyear, ccode, flag, fragment, democ, autoc, polity, polity2, durable, xrreg, xrcomp, xropen, xconst, parreg, parcomp, exrec, exconst, polcomp, prior, eprec, interim, bprec, post, change, d5, sf, regtrans, Beg, End, speccat.}
\item{FreedomHouse1: }{A dataset with 1885 observations and the following
47 variables: ID, COW_ID, Year, Label, Region, Edition, Status, PR rating, CL rating, A1, A2, A3, A, B1, B2, B3, B4, B, C1, C2, C3, C, Add Q, Add A, PR, D1, D2, D3, D4, D, E1, E2, E3, E, F1, F2, F3, F4, F, G1, G2, G3, G4, G, CL, Total, Territory.}
\item{FreedomHouse2: }{A dataset with 33264 observations and the following
7 variables: ID, COW_ID, Year, Label, Status, Value, Territory.}
\item{FreedomHouse3: }{A dataset with 3960 observations and the following
22 variables: ID, COW_ID, Year, Label, Region, Edition, Status, PR Rating, CL Rating, A, B, C, Add Q, Add A, PR, D, E, F, G, CL, Total, Territory.}
}
}
\usage{
regimes
}
\description{
regimes database documentation
}
\details{
\preformatted{## $Polity5
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             17574 
## Number of columns          35    
## _______________________          
## Column type frequency:           
##   character                5     
##   numeric                  30    
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate min max empty n_unique whitespace
## 1 ID                    0             1   8  12     0    17451          0
## 2 COW_ID                0             1   3   7     0      183          0
## 3 Label                 0             1   3  24     0      195          0
## 4 Beg                   0             1   4  10     0     1210          0
## 5 End                   0             1   4  10     0     1126          0
## 
## ── Variable type: numeric ──────────────────────────────────────────────────────
##    skim_variable n_missing complete_rate          mean           sd    p0
##  1 Year                  0        1         1941.           60.3     1776
##  2 p5                    0        1            0.311         0.463      0
##  3 cyear                 0        1      4041651.      2495259.     21776
##  4 ccode                 0        1          404.          250.         2
##  5 flag                  0        1            0.00319       0.0621     0
##  6 fragment          14128        0.196        0.229         0.711      0
##  7 democ               809        0.954        3.50          3.86       0
##  8 autoc               809        0.954        4.03          3.51       0
##  9 polity              809        0.954       -0.522         7.10     -10
## 10 polity2             268        0.985       -0.526         7.01     -10
## 11 durable            1264        0.928       20.9          24.6        0
## 12 xrreg               809        0.954        2.44          0.668      1
## 13 xrcomp              809        0.954        1.57          1.04       0
## 14 xropen              809        0.954        2.87          1.56       0
## 15 xconst              809        0.954        3.80          2.39       1
## 16 parreg              809        0.954        3.47          1.04       1
## 17 parcomp             809        0.954        2.64          1.47       0
## 18 exrec               183        0.990        0.821        17.7      -88
## 19 exconst             809        0.954        3.80          2.39       1
## 20 polcomp             126        0.993        1.25         17.8      -88
## 21 prior             16195        0.0785      -1.14          9.23     -88
## 22 eprec             16121        0.0827       1.10          5.16     -77
## 23 interim           16284        0.0734     -61.5          33.5      -88
## 24 bprec             16287        0.0732       1.42          1.18       1
## 25 post              16302        0.0724      -1.42         11.2      -88
## 26 change            15346        0.127       -5.20         46.3      -88
## 27 d5                16277        0.0738       1             0          1
## 28 sf                17276        0.0170       1             0          1
## 29 regtrans          15334        0.127       -5.85         45.1      -88
## 30 speccat           16765        0.0460     -77.7           9.23     -88
##         p25      p50      p75    p100
##  1    1894     1960     1992     2020
##  2       0        0        1        1
##  3 2001907. 3661924. 6301986. 9502018
##  4     200      366      630      950
##  5       0        0        0        2
##  6       0        0        0        3
##  7       0        1        7       10
##  8       0        4        7       10
##  9      -7       -3        7       10
## 10      -7       -3        7       10
## 11       4       13       29      170
## 12       2        3        3        3
## 13       1        1        3        3
## 14       1        4        4        4
## 15       1        3        7        7
## 16       3        4        4        5
## 17       1        3        4        5
## 18       2        4        8        8
## 19       1        3        7        7
## 20       1        6        7       10
## 21      -7       -3        5       99
## 22       1        1        1        9
## 23     -88      -77      -66        8
## 24       1        1        1        9
## 25      -7       -2        6       98
## 26     -11        0        3       99
## 27       1        1        1        1
## 28       1        1        1        1
## 29      -2        0        3       99
## 30     -88      -77      -66      -66
## 
## $FreedomHouse1
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             1885  
## Number of columns          47    
## _______________________          
## Column type frequency:           
##   character                7     
##   numeric                  40    
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate min max empty n_unique whitespace
## 1 ID                    0         1       7  12     0     1809          0
## 2 COW_ID               40         0.979   3   7     0      200          0
## 3 Year                  0         1       4   4     0        9          0
## 4 Label                 0         1       4  30     0      211          0
## 5 Region                0         1       3   8     0        6          0
## 6 Edition               0         1       4   4     0        9          0
## 7 Status                0         1       1   2     0        3          0
## 
## ── Variable type: numeric ──────────────────────────────────────────────────────
##    skim_variable n_missing complete_rate    mean     sd p0 p25 p50 p75 p100
##  1 PR rating             0         1      3.56    2.20   1   1   3   6    7
##  2 CL rating             0         1      3.45    1.93   1   2   3   5    7
##  3 A1                    0         1      2.43    1.61   0   1   3   4    4
##  4 A2                    0         1      2.53    1.52   0   1   3   4    4
##  5 A3                    0         1      2.40    1.41   0   1   3   4    4
##  6 A                     0         1      7.37    4.40   0   3   9  12   12
##  7 B1                    0         1      2.61    1.40   0   1   3   4    4
##  8 B2                    0         1      2.48    1.56   0   1   3   4    4
##  9 B3                    0         1      2.26    1.43   0   1   2   4    4
## 10 B4                    0         1      2.29    1.24   0   1   3   3    4
## 11 B                     0         1      9.64    5.36   0   5  11  15   16
## 12 C1                    0         1      2.27    1.51   0   1   3   4    4
## 13 C2                    0         1      1.86    1.17   0   1   2   3    4
## 14 C3                    0         1      1.97    1.29   0   1   2   3    4
## 15 C                     0         1      6.11    3.72   0   3   6  10   12
## 16 Add Q                 0         1      0.180   0.666  0   0   0   0    4
## 17 Add A               838         0.555  0.0516  0.349  0   0   0   0    3
## 18 PR                    0         1     23.0    13.4   -3  11  26  36   40
## 19 D1                    0         1      2.14    1.28   0   1   2   3    4
## 20 D2                    0         1      2.82    1.26   0   2   3   4    4
## 21 D3                    0         1      2.75    1.27   0   2   3   4    4
## 22 D4                    0         1      2.87    1.19   0   2   3   4    4
## 23 D                     0         1     10.6     4.64   0   7  12  15   16
## 24 E1                    0         1      2.41    1.45   0   1   3   4    4
## 25 E2                    0         1      2.50    1.42   0   1   3   4    4
## 26 E3                    0         1      2.39    1.26   0   1   2   3    4
## 27 E                     0         1      7.30    3.95   0   4   8  11   12
## 28 F1                    0         1      2.04    1.39   0   1   2   3    4
## 29 F2                    0         1      1.98    1.33   0   1   2   3    4
## 30 F3                    0         1      2.00    1.39   0   1   2   3    4
## 31 F4                    0         1      1.86    1.10   0   1   2   3    4
## 32 F                     0         1      7.88    4.94   0   4   7  12   16
## 33 G1                    0         1      2.64    1.25   0   2   3   4    4
## 34 G2                    0         1      2.37    1.10   0   2   2   3    4
## 35 G3                    0         1      2.29    1.05   0   2   2   3    4
## 36 G4                    0         1      2.01    1.06   0   1   2   3    4
## 37 G                     0         1      9.31    4.14   0   6   9  13   16
## 38 CL                    0         1     35.1    17.1    1  21  36  51   60
## 39 Total                 0         1     58.0    30.2   -1  31  62  87  100
## 40 Territory             0         1      0.0690  0.253  0   0   0   0    1
## 
## $FreedomHouse2
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             33264 
## Number of columns          7     
## _______________________          
## Column type frequency:           
##   character                6     
##   numeric                  1     
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate min max empty n_unique whitespace
## 1 ID                    0         1       7  12     0     9696          0
## 2 COW_ID             1440         0.957   3   7     0      201          0
## 3 Year                  0         1       4   4     0       48          0
## 4 Label                 0         1       4  46     0      230          0
## 5 Status                0         1       2   6     0        3          0
## 6 Value              5877         0.823   1   2     0       10          0
## 
## ── Variable type: numeric ──────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate  mean    sd p0 p25 p50 p75 p100
## 1 Territory             0             1 0.108 0.311  0   0   0   0    1
## 
## $FreedomHouse3
## ── Data Summary ────────────────────────
##                            Values
## Name                       X[[i]]
## Number of rows             3960  
## Number of columns          22    
## _______________________          
## Column type frequency:           
##   character                7     
##   numeric                  15    
## ________________________         
## Group variables            None  
## 
## ── Variable type: character ────────────────────────────────────────────────────
##   skim_variable n_missing complete_rate min max empty n_unique whitespace
## 1 ID                    0         1       7  12     0     3605          0
## 2 COW_ID               87         0.978   3   7     0      202          0
## 3 Year                  0         1       4   4     0       18          0
## 4 Label                 0         1       4  46     0      222          0
## 5 Region              621         0.843   3   8     0        6          0
## 6 Edition             621         0.843   4   4     0       16          0
## 7 Status              621         0.843   1   2     0        3          0
## 
## ── Variable type: numeric ──────────────────────────────────────────────────────
##    skim_variable n_missing complete_rate    mean     sd p0 p25 p50 p75 p100
##  1 PR Rating             3         0.999  6.61    9.06  -2   2   4   6   44
##  2 CL Rating             3         0.999  8.50   13.7    0   2   4   6   60
##  3 A                   621         0.843  7.41    4.38   0   3   9  12   12
##  4 B                   621         0.843  9.70    5.32   0   5  11  15   16
##  5 C                   621         0.843  6.20    3.73   0   3   6  10   12
##  6 Add Q               621         0.843  0.152   0.613  0   0   0   0    4
##  7 Add A              1459         0.632  0.0656  0.422  0   0   0   0    4
##  8 PR                  621         0.843 23.2    13.3   -3  11  26  36   40
##  9 D                   621         0.843 10.8     4.57   0   7  12  15   16
## 10 E                   621         0.843  7.45    3.90   0   4   8  11   12
## 11 F                   621         0.843  8.06    4.91   0   4   8  13   16
## 12 G                   621         0.843  9.38    4.11   0   6   9  13   16
## 13 CL                  621         0.843 35.7    16.9    0  22  36  52   60
## 14 Total                 3         0.999 59.0    29.8   -1  32  62  88  101
## 15 Territory             0         1      0.0687  0.253  0   0   0   0    1
}
}
\keyword{datasets}
