% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-regimes.R
\docType{data}
\name{regimes}
\alias{regimes}
\title{regimes database documentation}
\format{
The regimes database is a list that contains the
following 4 datasets: Polity5, FHaggregate, FHstatus, FHfull.
FHaggregate, FHstatus, and FHfull datasets are all part of
the FreedomHouse data that have been imported into the package
as three datasets as they are provided as three distinct Excel files.
The FHaggregate dataset contains aggregate scores from the 2003-2021 edition
and the FHfull dataset contains detailed scores from the 2013-2021 edition.
The FHstatus dataset contains only data on the Freedom House status of
countries and territories.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{Polity5: }{A dataset with 17574 observations and the following
35 variables: ID, cowID, Year, Label, p5, cyear, ccode, flag, fragment, democ, autoc, polity, polity2, durable, xrreg, xrcomp, xropen, xconst, parreg, parcomp, exrec, exconst, polcomp, prior, eprec, interim, bprec, post, change, d5, sf, regtrans, Beg, End, speccat.}
\item{FHaggregate: }{A dataset with 3960 observations and the following
22 variables: ID, cowID, Year, Label, Region, Edition, Status, PR rating, CL rating, A, B, C, Add Q, Add A, PR, D, E, F, G, CL, Total, Territory.}
\item{FHstatus: }{A dataset with 11088 observations and the following
8 variables: ID, cowID, Year, Label, PR rating, CL rating, Status, Territory.}
\item{FHfull: }{A dataset with 1885 observations and the following
47 variables: ID, cowID, Year, Label, Region, Edition, Status, PR rating, CL rating, A1, A2, A3, A, B1, B2, B3, B4, B, C1, C2, C3, C, Add Q, Add A, PR, D1, D2, D3, D4, D, E1, E2, E3, E, F1, F2, F3, F4, F, G1, G2, G3, G4, G, CL, Total, Territory.}
}
}
\usage{
regimes
}
\description{
regimes database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $Polity5
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |      ID       |  character  |           17574|     0     |       0       |
#> |     cowID     |  character  |           17574|     0     |       0       |
#> |     Year      |    mdate    |           17574|     0     |       0       |
#> |     Label     |  character  |           17574|     0     |       0       |
#> |      p5       |   numeric   |           17574|     0     |       0       |
#> |     cyear     |   numeric   |           17574|     0     |       0       |
#> |     ccode     |   numeric   |           17574|     0     |       0       |
#> |     flag      |   numeric   |           17574|     0     |       0       |
#> |   fragment    |   numeric   |           17574|   14128   |     80.39     |
#> |     democ     |   numeric   |           17574|    809    |      4.6      |
#> |     autoc     |   numeric   |           17574|    809    |      4.6      |
#> |    polity     |   numeric   |           17574|    809    |      4.6      |
#> |    polity2    |   numeric   |           17574|    268    |     1.52      |
#> |    durable    |   numeric   |           17574|   1264    |     7.19      |
#> |     xrreg     |   numeric   |           17574|    809    |      4.6      |
#> |    xrcomp     |   numeric   |           17574|    809    |      4.6      |
#> |    xropen     |   numeric   |           17574|    809    |      4.6      |
#> |    xconst     |   numeric   |           17574|    809    |      4.6      |
#> |    parreg     |   numeric   |           17574|    809    |      4.6      |
#> |    parcomp    |   numeric   |           17574|    809    |      4.6      |
#> |     exrec     |   numeric   |           17574|    183    |     1.04      |
#> |    exconst    |   numeric   |           17574|    809    |      4.6      |
#> |    polcomp    |   numeric   |           17574|    126    |     0.72      |
#> |     prior     |   numeric   |           17574|   16195   |     92.15     |
#> |     eprec     |   numeric   |           17574|   16121   |     91.73     |
#> |    interim    |   numeric   |           17574|   16284   |     92.66     |
#> |     bprec     |   numeric   |           17574|   16287   |     92.68     |
#> |     post      |   numeric   |           17574|   16302   |     92.76     |
#> |    change     |   numeric   |           17574|   15346   |     87.32     |
#> |      d5       |   numeric   |           17574|   16277   |     92.62     |
#> |      sf       |   numeric   |           17574|   17276   |     98.3      |
#> |   regtrans    |   numeric   |           17574|   15334   |     87.25     |
#> |      Beg      |    mdate    |           17574|     0     |       0       |
#> |      End      |    mdate    |           17574|     0     |       0       |
#> |    speccat    |   numeric   |           17574|   16765   |     95.4      |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $FHaggregate
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |      ID       |  character  |            3960|     0     |       0       |
#> |     cowID     |  character  |            3960|    87     |      2.2      |
#> |     Year      |    mdate    |            3960|     0     |       0       |
#> |     Label     |  character  |            3960|     0     |       0       |
#> |    Region     |  character  |            3960|    621    |     15.68     |
#> |    Edition    |    mdate    |            3960|    621    |     15.68     |
#> |    Status     |  character  |            3960|    621    |     15.68     |
#> |   PR rating   |   numeric   |            3960|     3     |     0.08      |
#> |   CL rating   |   numeric   |            3960|     3     |     0.08      |
#> |       A       |   numeric   |            3960|    621    |     15.68     |
#> |       B       |   numeric   |            3960|    621    |     15.68     |
#> |       C       |   numeric   |            3960|    621    |     15.68     |
#> |     Add Q     |   numeric   |            3960|    621    |     15.68     |
#> |     Add A     |   numeric   |            3960|   1459    |     36.84     |
#> |      PR       |   numeric   |            3960|    621    |     15.68     |
#> |       D       |   numeric   |            3960|    621    |     15.68     |
#> |       E       |   numeric   |            3960|    621    |     15.68     |
#> |       F       |   numeric   |            3960|    621    |     15.68     |
#> |       G       |   numeric   |            3960|    621    |     15.68     |
#> |      CL       |   numeric   |            3960|    621    |     15.68     |
#> |     Total     |   numeric   |            3960|     3     |     0.08      |
#> |   Territory   |   numeric   |            3960|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $FHstatus
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |      ID       |  character  |           11088|     0     |       0       |
#> |     cowID     |  character  |           11088|    480    |     4.33      |
#> |     Year      |    mdate    |           11088|     0     |       0       |
#> |     Label     |  character  |           11088|     0     |       0       |
#> |   PR rating   |   numeric   |           11088|   1959    |     17.67     |
#> |   CL rating   |   numeric   |           11088|   1959    |     17.67     |
#> |    Status     |  character  |           11088|   1959    |     17.67     |
#> |   Territory   |   numeric   |           11088|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $FHfull
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |      ID       |  character  |            1885|     0     |       0       |
#> |     cowID     |  character  |            1885|    40     |     2.12      |
#> |     Year      |    mdate    |            1885|     0     |       0       |
#> |     Label     |  character  |            1885|     0     |       0       |
#> |    Region     |  character  |            1885|     0     |       0       |
#> |    Edition    |    mdate    |            1885|     0     |       0       |
#> |    Status     |  character  |            1885|     0     |       0       |
#> |   PR rating   |   numeric   |            1885|     0     |       0       |
#> |   CL rating   |   numeric   |            1885|     0     |       0       |
#> |      A1       |   numeric   |            1885|     0     |       0       |
#> |      A2       |   numeric   |            1885|     0     |       0       |
#> |      A3       |   numeric   |            1885|     0     |       0       |
#> |       A       |   numeric   |            1885|     0     |       0       |
#> |      B1       |   numeric   |            1885|     0     |       0       |
#> |      B2       |   numeric   |            1885|     0     |       0       |
#> |      B3       |   numeric   |            1885|     0     |       0       |
#> |      B4       |   numeric   |            1885|     0     |       0       |
#> |       B       |   numeric   |            1885|     0     |       0       |
#> |      C1       |   numeric   |            1885|     0     |       0       |
#> |      C2       |   numeric   |            1885|     0     |       0       |
#> |      C3       |   numeric   |            1885|     0     |       0       |
#> |       C       |   numeric   |            1885|     0     |       0       |
#> |     Add Q     |   numeric   |            1885|     0     |       0       |
#> |     Add A     |   numeric   |            1885|    838    |     44.46     |
#> |      PR       |   numeric   |            1885|     0     |       0       |
#> |      D1       |   numeric   |            1885|     0     |       0       |
#> |      D2       |   numeric   |            1885|     0     |       0       |
#> |      D3       |   numeric   |            1885|     0     |       0       |
#> |      D4       |   numeric   |            1885|     0     |       0       |
#> |       D       |   numeric   |            1885|     0     |       0       |
#> |      E1       |   numeric   |            1885|     0     |       0       |
#> |      E2       |   numeric   |            1885|     0     |       0       |
#> |      E3       |   numeric   |            1885|     0     |       0       |
#> |       E       |   numeric   |            1885|     0     |       0       |
#> |      F1       |   numeric   |            1885|     0     |       0       |
#> |      F2       |   numeric   |            1885|     0     |       0       |
#> |      F3       |   numeric   |            1885|     0     |       0       |
#> |      F4       |   numeric   |            1885|     0     |       0       |
#> |       F       |   numeric   |            1885|     0     |       0       |
#> |      G1       |   numeric   |            1885|     0     |       0       |
#> |      G2       |   numeric   |            1885|     0     |       0       |
#> |      G3       |   numeric   |            1885|     0     |       0       |
#> |      G4       |   numeric   |            1885|     0     |       0       |
#> |       G       |   numeric   |            1885|     0     |       0       |
#> |      CL       |   numeric   |            1885|     0     |       0       |
#> |     Total     |   numeric   |            1885|     0     |       0       |
#> |   Territory   |   numeric   |            1885|     0     |       0       |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
