% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-states.R
\docType{data}
\name{states}
\alias{states}
\title{States datacube}
\format{
\describe{
\item{ISD: }{A dataset with 499
observations and 36 variables:
stateID, StateName, Begin, End, StateNameAlt, Latitude, Longitude, StartType, EndType, cowID, cowNR, ISD_Category, Region, Start_Am, EStart_Am, Declare, DecDate, Population, ..., VioEnd, and VioEnd_Am.}
\item{HUGGO: }{A dataset with
observations and  variables:
.}
\item{GW: }{A dataset with 216
observations and 7 variables:
stateID, StateName, Begin, End, StateNameAlt, cowID, and cowNR.}
}
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
}
\source{
\itemize{
\item Griffiths, Ryan D., and Charles R. Butcher. 'Introducing the international system(s) dataset (ISD), 1816-2011'. International Interactions 39.5 (2013), pp. 748-768.
\item Gleditsch, Kristian S., and Michael D. Ward. 'Interstate system membership: A revised list of the independent states since 1816'. International Interactions 25.4 (1999), pp. 393-413.
\item Hand-coded data by the GGO team
}
}
\usage{
states
}
\description{
The \code{manystates::states} datacube is a list containing 3 datasets: ISD, GW, and GGO.
It is a work-in-progress, so please let us know if you have any comments or suggestions.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $ISD
#> ---------------------------------------------------------
#> |   Variable   |  Class  |  Obs  |  Missing  |  Miss \%  |
#> ---------------------------------------------------------
#> |stateID       |character|    499|          0|         0|
#> |StateName     |character|    499|          0|         0|
#> |Begin         |mdate    |    282|        217|     43.49|
#> |End           |mdate    |    499|          0|         0|
#> |StateNameAlt  |character|    210|        289|     57.92|
#> |Latitude      |character|    343|        156|     31.26|
#> |Longitude     |character|    343|        156|     31.26|
#> |StartType     |numeric  |    337|        162|     32.46|
#> |EndType       |numeric  |    296|        203|     40.68|
#> |cowID         |character|    499|          0|         0|
#> |cowNR         |numeric  |    499|          0|         0|
#> |ISD_Category  |numeric  |    497|          2|       0.4|
#> |Region        |numeric  |    499|          0|         0|
#> |Start_Am      |numeric  |    499|          0|         0|
#> |EStart_Am     |numeric  |    232|        267|     53.51|
#> |Declare       |numeric  |    315|        184|     36.87|
#> |DecDate       |character|     72|        427|     85.57|
#> |Population    |character|    319|        180|     36.07|
#> |PopDate       |numeric  |    316|        183|     36.67|
#> |PopAm         |numeric  |    339|        160|     32.06|
#> |PopulationHigh|numeric  |    139|        360|     72.14|
#> |PopulationLow |numeric  |    108|        391|     78.36|
#> |StartType_Am  |numeric  |    339|        160|     32.06|
#> |StartSettle   |numeric  |    320|        179|     35.87|
#> |End_Am        |numeric  |    499|          0|         0|
#> |EndType_Am    |numeric  |    294|        205|     41.08|
#> |EndSettle     |numeric  |    282|        217|     43.49|
#> |Sovereignty_Am|numeric  |    499|          0|         0|
#> |EuroDip       |numeric  |    331|        168|     33.67|
#> |Borders       |numeric  |    332|        167|     33.47|
#> |Borders_Am    |numeric  |    342|        157|     31.46|
#> |Capital       |character|    284|        215|     43.09|
#> |VioStart      |numeric  |    318|        181|     36.27|
#> |VioStart_Am   |numeric  |    327|        172|     34.47|
#> |VioEnd        |numeric  |    292|        207|     41.48|
#> |VioEnd_Am     |numeric  |    297|        202|     40.48|
#> ---------------------------------------------------------
#> 
#> 
#> $GW
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |stateID     |character|    216|          0|         0|
#> |StateName   |character|    216|          0|         0|
#> |Begin       |mdate    |    216|          0|         0|
#> |End         |mdate    |    216|          0|         0|
#> |StateNameAlt|character|     20|        196|     90.74|
#> |cowID       |character|    216|          0|         0|
#> |cowNR       |character|    216|          0|         0|
#> -------------------------------------------------------
#> 
#> 
#> $GGO
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |stateID     |character|    407|          0|         0|
#> |StateName   |character|    407|          0|         0|
#> |Capital     |character|    406|          1|      0.25|
#> |Begin       |mdate    |    406|          1|      0.25|
#> |End         |mdate    |    407|          0|         0|
#> |Latitude    |numeric  |    407|          0|         0|
#> |Longitude   |numeric  |    407|          0|         0|
#> |Region      |character|    407|          0|         0|
#> |StateNameAlt|character|     55|        352|     86.49|
#> |CapitalAlt  |character|      7|        400|     98.28|
#> |RatProcedure|character|     56|        351|     86.24|
#> |Coder       |character|     90|        317|     77.89|
#> |Comments    |character|     76|        331|     81.33|
#> |Source      |character|    110|        297|     72.97|
#> -------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{Mapping}{
\tabular{lll}{
   \emph{GGO} \tab \emph{GW} \tab \emph{ISD} \cr
   stateID \tab  \tab  \cr
   Begin \tab Start \tab Start \cr
   End \tab Finish \tab Finish \cr
   StateName \tab Name of State \tab State.Name \cr
   cowID \tab Cow ID \tab COW.ID \cr
   cowNR \tab Cow NR. \tab COW.Nr \cr
}
}

\keyword{datasets}
