% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\docType{data}
\name{states}
\alias{states}
\title{Data on the entry and exit of sovereign states to the interstate system containing
the following datasets:}
\format{
This dataset contains 243 observations and 5 variables:
\describe{
\item{COW_Nr}{Numerical COW identifier}
\item{ID}{# letter abbreviation of State name}
\item{Beg}{Beginning	date	of	state	tenure}
\item{End}{End date of state tenure}
\item{Label}{Primary COW State name}
}
}
\source{
\url{https://correlatesofwar.org/data-sets/state-system-membership}
}
\usage{
states
}
\description{
Data on the entry and exit of sovereign states to the interstate system containing
the following datasets:
}
\keyword{datasets}
