% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-states.R
\docType{data}
\name{states}
\alias{states}
\title{states database documentation}
\format{
The states database is a list that contains the
following 7 datasets: GW, ISD, COW, ICOW, ICOW_COL, RATRULES, HUGGO_STATES.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{GW: }{A dataset with 216 observations and the following
5 variables: cowID, Beg, End, cowNR, StateName.}
\item{ISD: }{A dataset with 362 observations and the following
7 variables: cowID, Beg, End, cowNR, StateName, Micro, NewState.}
\item{COW: }{A dataset with 243 observations and the following
5 variables: cowID, Beg, End, cowNR, StateName.}
\item{ICOW: }{A dataset with 217 observations and the following
15 variables: cowID, StateName, ColRuler, IndFrom, IndDate, IndViol, IndType,
SecFrom, SecDate, SecViol, Into, IntoDate, COWsys, GWsys, Notes.}
\item{ICOW_COL: }{A dataset with 221 observations and the following
9 variables: cowID, cowNR, StateName, Beg, cowID_Origin, IndepType, Beg_COW,
Beg_GW, Beg_Polity2.}
\item{RATRULES: }{A dataset with 177 observations and the following
3 variables: stateID, StateName, RatProcedure.}
\item{HUGGO_STATES: }{A dataset with 574 observations and the following
12 variables: stateID, StateName, Capital, Beg, End, Latitude, Longitude,
Area, Region, RatProcedure, Constitutional Description, Source.}
}
}
\usage{
states
}
\description{
states database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $GW
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |     cowID     |  character  |             216|     0     |       0       |
#> |      Beg      |    mdate    |             216|     0     |       0       |
#> |      End      |    mdate    |             216|     0     |       0       |
#> |     cowNR     |  character  |             216|     0     |       0       |
#> |   StateName   |  character  |             216|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $ISD
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |     cowID     |  character  |             362|     0     |       0       |
#> |      Beg      |    mdate    |             362|     0     |       0       |
#> |      End      |    mdate    |             362|     0     |       0       |
#> |     cowNR     |  character  |             362|     0     |       0       |
#> |   StateName   |  character  |             362|     0     |       0       |
#> |     Micro     |  character  |             362|     0     |       0       |
#> |   NewState    |  character  |             362|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $COW
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |     cowID     |  character  |             243|     0     |       0       |
#> |      Beg      |    mdate    |             243|     0     |       0       |
#> |      End      |    mdate    |             243|     0     |       0       |
#> |     cowNR     |  character  |             243|     0     |       0       |
#> |   StateName   |  character  |             243|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $ICOW
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |     cowID     |  character  |             217|     0     |       0       |
#> |   StateName   |  character  |             217|     0     |       0       |
#> |   ColRuler    |   numeric   |             217|    46     |     21.2      |
#> |    IndFrom    |   numeric   |             217|    39     |     17.97     |
#> |    IndDate    |    mdate    |             217|     0     |       0       |
#> |    IndViol    |   numeric   |             217|     0     |       0       |
#> |    IndType    |   numeric   |             217|     0     |       0       |
#> |    SecFrom    |   numeric   |             217|    179    |     82.49     |
#> |    SecDate    |    mdate    |             217|    179    |     82.49     |
#> |    SecViol    |   numeric   |             217|    179    |     82.49     |
#> |     Into      |   numeric   |             217|    198    |     91.24     |
#> |   IntoDate    |    mdate    |             217|    198    |     91.24     |
#> |    COWsys     |    mdate    |             217|     0     |       0       |
#> |     GWsys     |    mdate    |             217|    22     |     10.14     |
#> |     Notes     |  character  |             217|    58     |     26.73     |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $ICOW_COL
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |     cowID     |  character  |             221|     2     |      0.9      |
#> |     cowNR     |   numeric   |             221|     0     |       0       |
#> |   StateName   |  character  |             221|     0     |       0       |
#> |      Beg      |    mdate    |             221|     0     |       0       |
#> | cowID_Origin  |   numeric   |             221|     0     |       0       |
#> |   IndepType   |   numeric   |             221|     0     |       0       |
#> |    Beg_COW    |    mdate    |             221|     0     |       0       |
#> |    Beg_GW     |    mdate    |             221|    23     |     10.41     |
#> |  Beg_Polity2  |    mdate    |             221|    74     |     33.48     |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $RATRULES
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    stateID    |  character  |             177|     0     |       0       |
#> |   StateName   |  character  |             177|     0     |       0       |
#> | RatProcedure  |   numeric   |             177|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO_STATES
#> ---------------------------------------------------------------------------------------
#> |       Column Name        |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ---------------------------------------------------------------------------------------
#> |         stateID          |  character  |             574|     0     |       0       |
#> |        StateName         |  character  |             574|    214    |     37.28     |
#> |         Capital          |  character  |             574|    215    |     37.46     |
#> |           Beg            |    mdate    |             574|    257    |     44.77     |
#> |           End            |    mdate    |             574|    257    |     44.77     |
#> |         Latitude         |  character  |             574|    234    |     40.77     |
#> |        Longitude         |  character  |             574|    234    |     40.77     |
#> |           Area           |  character  |             574|    448    |     78.05     |
#> |          Region          |  character  |             574|    356    |     62.02     |
#> |       RatProcedure       |  character  |             574|    542    |     94.43     |
#> |Constitutional Description|  character  |             574|    542    |     94.43     |
#> |          Source          |  character  |             574|    533    |     92.86     |
#> ---------------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
