% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qStates-states.R
\docType{data}
\name{states}
\alias{states}
\title{states database documentation}
\format{
The states database is a list that contains the
following 3 datasets: GW, ISD, COW.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()}
functions.
\describe{
\item{GW: }{A dataset with 216 observations and the
following 5 variables: ID, Beg, End, COW_Nr, Label.}
\item{ISD: }{A dataset with 362 observations and the
following 7 variables: ID, Beg, End, COW_Nr, Label,
Micro, New.State.}
\item{COW: }{A dataset with 243 observations and the
following 5 variables: ID, Beg, End, COW_Nr, Label.}
}
}
\usage{
states
}
\description{
states database documentation
}
\keyword{datasets}
