% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\docType{data}
\name{states}
\alias{states}
\title{Data on the entry and exit of sovereign states to the interstate system}
\format{
This database contains 3 datasets. 216 observations and 5 variables:
\describe{
\item{COW_Nr}{Reference number for country according to COW project}
\item{ID}{Three letter country ID according to COW project}
\item{Beg}{Beginning	date	of	state	tenure}
\item{End}{End date of state tenure}
\item{Label}{Primary COW State name}
}
}
\source{
\url{http://ksgleditsch.com/statelist.html,http://ksgleditsch.com/data/iisyst_casedesc.pdf}
}
\usage{
states
}
\description{
This dataset contains a revised and updated version of Gleditsch and Ward (1999) list of independent states.
Please refer to the original article and the revised codebook for a detailed
discussion of the coding rules, sources, general criteria  and issues in assessing
effective independence. The data contain some changes over the table in the Gleditsch and Ward (1999) article.
These are documented in the revised codebook. The additional (tentative) list of microstates and dates
provided by the authors in the supplementary materials has not been added here.
}
\keyword{datasets}
