% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\docType{data}
\name{states}
\alias{states}
\title{Data on the entry and exit of sovereign states to the interstate system containing
the following datasets:}
\format{
This dataset contains 243 observations and 5 variables:
\describe{
\item{COW_Nr}{Numerical COW identifier}
\item{ID}{# letter abbreviation of State name}
\item{Beg}{Beginning	date	of	state	tenure}
\item{End}{End date of state tenure}
\item{Label}{Primary COW State name}
}

This database contains 3 datasets. 216 observations and 5 variables:
\describe{
\item{COW_Nr}{Reference number for country according to COW project}
\item{ID}{Three letter country ID according to COW project}
\item{Beg}{Beginning	date	of	state	tenure}
\item{End}{End date of state tenure}
\item{Label}{Primary COW State name}
}

This dataset contains  363 observations and 7 variables:
\describe{
\item{COW_Nr}{Reference number for country according to COW project}
\item{ID}{Three letter country ID according to COW project}
\item{Beg}{Date of beginning of State tenure}
\item{End}{Date of ending of State tenure}
\item{Label}{Name of the state}
\item{Micro}{Whether or not the State is considered a micro state, dummy variable}
\item{New.State}{Whether or not the State appeared in COW or GW datasets before the date of beginning of the state tenure, dummy variable}
}
}
\source{
\url{https://correlatesofwar.org/data-sets/state-system-membership}

"GW"

This dataset contains a revised and updated version of Gleditsch and Ward (1999) list of independent states.
Please refer to the original article and the revised codebook for a detailed
discussion of the coding rules, sources, general criteria  and issues in assessing
effective independence. The data contain some changes over the table in the Gleditsch and Ward (1999) article.
These are documented in the revised codebook. The additional (tentative) list of microstates and dates
provided by the authors in the supplementary materials has not been added here.

\url{http://ksgleditsch.com/statelist.html,http://ksgleditsch.com/data/iisyst_casedesc.pdf}

"ISD"

The International Systems Dataset adds 96 completely new states. Some of these states were earlier
incarnations of states that would exist later, typically during the 20th century (e.g. Bhutan), but
the majority did not reappear. See the Griffiths and Butcher(2013) article for a conceptual
discussion and the codebook for details on coding rules and cases here
(https://static1.squarespace.com/static/54eccfa0e4b08d8eee5174af/t/54ede030e4b0a0f14faa8f84/1424875568381/ISD+Codebook_version1.pdf)

\url{https://www.tandfonline.com/doi/full/10.1080/03050629.2013.834259}
}
\usage{
states
}
\description{
"COW"
}
\details{
This dataset contains list of states in the international system as updated and distributed by
the Correlates of War Project (COW). The dates range from 1816 until 2016.
}
\keyword{datasets}
