% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-states.R
\docType{data}
\name{states}
\alias{states}
\title{states database documentation}
\format{
The states database is a list that contains the
following 4 datasets: GW, ISD, COW, ICOW.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{GW: }{A dataset with 216 observations and the following
5 variables: COW_ID, Beg, End, COW_Nr, Label.}
\item{ISD: }{A dataset with 362 observations and the following
7 variables: COW_ID, Beg, End, COW_Nr, Label, Micro, New.State.}
\item{COW: }{A dataset with 243 observations and the following
5 variables: COW_ID, Beg, End, COW_Nr, Label.}
\item{ICOW: }{A dataset with 217 observations and the following
15 variables: COW_ID, Label, ColRuler, IndFrom, IndDate, IndViol, IndType, SecFrom, SecDate, SecViol, Into, IntoDate, COWsys, GWsys, Notes.}
}
}
\usage{
states
}
\description{
states database documentation
}
\details{
\preformatted{## $GW
## -- Data Summary ------------------------
##                            Values
## Name                       X[[i]]
## Number of rows             216   
## Number of columns          5     
## _______________________          
## Column type frequency:           
##   character                5     
## ________________________         
## Group variables            None  
## 
## -- Variable type: character ----------------------------------------------------
## # A tibble: 5 x 8
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## * <chr>             <int>         <dbl> <int> <int> <int>    <int>      <int>
## 1 COW_ID                0             1     3     3     0      202          0
## 2 Beg                   0             1    10    12     0      153          0
## 3 End                   0             1    10    12     0       30          0
## 4 COW_Nr                0             1     1     3     0      202          0
## 5 Label                 0             1     4    39     0      203          0
## 
## $ISD
## -- Data Summary ------------------------
##                            Values
## Name                       X[[i]]
## Number of rows             362   
## Number of columns          7     
## _______________________          
## Column type frequency:           
##   character                7     
## ________________________         
## Group variables            None  
## 
## -- Variable type: character ----------------------------------------------------
## # A tibble: 7 x 8
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## * <chr>             <int>         <dbl> <int> <int> <int>    <int>      <int>
## 1 COW_ID                0             1     3     3     0      310          0
## 2 Beg                   0             1    10    12     0      201          0
## 3 End                   0             1    10    12     0      134          0
## 4 COW_Nr                0             1     1     4     0      315          0
## 5 Label                 0             1     3    35     0      319          0
## 6 Micro                 0             1     1     1     0        2          0
## 7 New.State             0             1     1     1     0        2          0
## 
## $COW
## -- Data Summary ------------------------
##                            Values
## Name                       X[[i]]
## Number of rows             243   
## Number of columns          5     
## _______________________          
## Column type frequency:           
##   character                5     
## ________________________         
## Group variables            None  
## 
## -- Variable type: character ----------------------------------------------------
## # A tibble: 5 x 8
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## * <chr>             <int>         <dbl> <int> <int> <int>    <int>      <int>
## 1 COW_ID                0             1     3     3     0      217          0
## 2 Beg                   0             1    10    12     0      191          0
## 3 End                   0             1    10    12     0       39          0
## 4 COW_Nr                0             1     1     3     0      217          0
## 5 Label                 0             1     4    32     0      217          0
## 
## $ICOW
## -- Data Summary ------------------------
##                            Values
## Name                       X[[i]]
## Number of rows             217   
## Number of columns          15    
## _______________________          
## Column type frequency:           
##   character                8     
##   numeric                  7     
## ________________________         
## Group variables            None  
## 
## -- Variable type: character ----------------------------------------------------
## # A tibble: 8 x 8
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## * <chr>             <int>         <dbl> <int> <int> <int>    <int>      <int>
## 1 COW_ID                0        1          3     3     0      217          0
## 2 Label                 0        1          4    36     0      217          0
## 3 IndDate               0        1          6     8     0      175          0
## 4 SecDate             179        0.175      7     7     0       31          0
## 5 IntoDate            198        0.0876     7     7     0       11          0
## 6 COWsys                0        1          7     7     0      142          0
## 7 GWsys                22        0.899      7     7     0      122          0
## 8 Notes                58        0.733     25   239     0      138          0
## 
## -- Variable type: numeric ------------------------------------------------------
## # A tibble: 7 x 10
##   skim_variable n_missing complete_rate    mean      sd    p0   p25   p50   p75
## * <chr>             <int>         <dbl>   <dbl>   <dbl> <dbl> <dbl> <dbl> <dbl>
## 1 ColRuler             46        0.788  293.    172.        2  200    220   300
## 2 IndFrom              39        0.820  289.    185.        2  200    220   345
## 3 IndViol               0        1        0.424   0.495     0    0      0     1
## 4 IndType               0        1        2.15    0.774     1    2      2     2
## 5 SecFrom             179        0.175  313.    200.       41  141.   315   365
## 6 SecViol             179        0.175    0.605   0.495     0    0      1     1
## 7 Into                198        0.0876 361     175.      255  255    255   325
##    p100
## * <dbl>
## 1   920
## 2   920
## 3     1
## 4     4
## 5   820
## 6     1
## 7   816
}
}
\keyword{datasets}
