% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-states.R
\docType{data}
\name{states}
\alias{states}
\title{states datacube documentation}
\format{
The states datacube is a list that contains the
following 8 datasets: GW, ISD, COW, ICOW, RATRULES, and HUGGO_STATES.
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
\describe{
\item{GW: }{A dataset with 216 observations and the following
6 variables: cowID, Begin, End, cowNR, StateName, stateID.}
\item{ISD: }{A dataset with 362 observations and the following
8 variables: cowID, Begin, End, cowNR, StateName, Micro, NewState, stateID.}
\item{COW: }{A dataset with 243 observations and the following
6 variables: cowID, Begin, End, cowNR, StateName, stateID.}
\item{ICOW: }{A dataset with 217 observations and the following
16 variables: stateID, StateName, Begin, ColRuler, IndFrom, IndViol, IndType,
SecFrom, SecDate, SecViol, Into, IntoDate, COWsys, GWsys, Notes, cowID.}
\item{RATRULES: }{A dataset with 177 observations and the following
3 variables: stateID, StateName, RatProcedure.}
\item{HUGGO_STATES: }{A dataset with 470 observations and the following
14 variables: stateID, StateName, Capital, Begin, End, StateName2, Capital2,
Latitude, Longitude, Area, Region, RatProcedure, Constitutional Description,
Source_rat.}
}
\source{
\itemize{
\item{GW: }{
K. S. Gleditsch and M. D. Ward. “Interstate system membership: A revised list of the independent states since 1816”.
\emph{International Interactions} 25.4 (1999), pp. 393-413.
\url{http://ksgleditsch.com/data-4.html}
}
\item{ISD: }{
R. D. Griffiths and C. R. Butcher. “Introducing the international system (s) dataset (ISD), 1816-2011”.
\emph{International Interactions} 39.5 (2013), pp. 748-768.
\url{http://www.ryan-griffiths.com/data}
}
\item{COW: }{
Correlates of War Project. \emph{State System Membership List, v2016.}
\url{http://correlatesofwar.org}. Accessed: 2021-01-25.
\url{https://correlatesofwar.org/data-sets/state-system-membership}
}
\item{ICOW: }{
P. R. Hensel. (2018). \emph{ICOW Colonial History Data Set, version 1.1.}
\url{http://www.paulhensel.org/icowcol.html}.Accessed: 2021-12-23.
\url{http://www.paulhensel.org/icowcol.html}
}
\item{RATRULES: }{
B. A. Simmons. \emph{Mobilizing for Human Rights: International Law in Domestic Politics}.
Cambridge University Press, 2009.
\url{https://doi.org/10.1017/CBO9780511811340}
}
\item{HUGGO_STATES: }{
J. Hollway. manystates: States for manydata. 2021.
Hand-coded by the GGO team.}
}
\usage{
states
}
\description{
states datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $GW
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |     cowID     |  character  |             216|     0     |       0       |
#> |     Begin     |    mdate    |             216|     0     |       0       |
#> |      End      |    mdate    |             216|     0     |       0       |
#> |     cowNR     |  character  |             216|     0     |       0       |
#> |   StateName   |  character  |             216|     0     |       0       |
#> |    stateID    |  character  |             216|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $ISD
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |     cowID     |  character  |             362|     0     |       0       |
#> |     Begin     |    mdate    |             362|     0     |       0       |
#> |      End      |    mdate    |             362|     0     |       0       |
#> |     cowNR     |  character  |             362|     0     |       0       |
#> |   StateName   |  character  |             362|     0     |       0       |
#> |     Micro     |  character  |             362|     0     |       0       |
#> |   NewState    |  character  |             362|     0     |       0       |
#> |    stateID    |  character  |             362|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $COW
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |     cowID     |  character  |             243|     0     |       0       |
#> |     Begin     |    mdate    |             243|     0     |       0       |
#> |      End      |    mdate    |             243|     0     |       0       |
#> |     cowNR     |  character  |             243|     0     |       0       |
#> |   StateName   |  character  |             243|     0     |       0       |
#> |    stateID    |  character  |             243|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $ICOW
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    stateID    |  character  |             217|     0     |       0       |
#> |   StateName   |  character  |             217|     0     |       0       |
#> |     Begin     |    mdate    |             217|     0     |       0       |
#> |   ColRuler    |  character  |             217|    46     |     21.2      |
#> |    IndFrom    |  character  |             217|    39     |     17.97     |
#> |    IndViol    |  character  |             217|     0     |       0       |
#> |    IndType    |  character  |             217|     0     |       0       |
#> |    SecFrom    |  character  |             217|    179    |     82.49     |
#> |    SecDate    |    mdate    |             217|    179    |     82.49     |
#> |    SecViol    |  character  |             217|    179    |     82.49     |
#> |     Into      |  character  |             217|    198    |     91.24     |
#> |   IntoDate    |    mdate    |             217|    198    |     91.24     |
#> |    COWsys     |    mdate    |             217|     0     |       0       |
#> |     GWsys     |    mdate    |             217|    22     |     10.14     |
#> |     Notes     |  character  |             217|    58     |     26.73     |
#> |     cowID     |  character  |             217|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $RATRULES
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    stateID    |  character  |             177|     0     |       0       |
#> |   StateName   |  character  |             177|     0     |       0       |
#> | RatProcedure  |  character  |             177|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO_STATES
#> ---------------------------------------------------------------------------------------
#> |       Column Name        |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ---------------------------------------------------------------------------------------
#> |         stateID          |  character  |             470|     0     |       0       |
#> |        StateName         |  character  |             470|     2     |     0.43      |
#> |         Capital          |  character  |             470|     3     |     0.64      |
#> |          Begin           |    mdate    |             470|    150    |     31.91     |
#> |           End            |    mdate    |             470|    150    |     31.91     |
#> |        StateName2        |  character  |             470|    464    |     98.72     |
#> |         Capital2         |  character  |             470|    469    |     99.79     |
#> |         Latitude         |  character  |             470|     3     |     0.64      |
#> |        Longitude         |  character  |             470|     3     |     0.64      |
#> |           Area           |  character  |             470|    259    |     55.11     |
#> |          Region          |  character  |             470|    259    |     55.11     |
#> |       RatProcedure       |  character  |             470|    440    |     93.62     |
#> |Constitutional Description|  character  |             470|    440    |     93.62     |
#> |        Source_rat        |  character  |             470|    432    |     91.91     |
#> ---------------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{Mapping}{
\tabular{llllll}{
   \emph{manystates} \tab \emph{GW} \tab \emph{ISD} \tab \emph{COW} \tab \emph{ICOW} \tab \emph{RATRULES} \cr
   stateID \tab  \tab  \tab  \tab  \tab StatID \cr
   cowID \tab Cow ID \tab COW.ID \tab stateabb \tab State \tab  \cr
   Begin \tab Start \tab Start \tab styear,stmonth,stday \tab IndDate \tab  \cr
   End \tab Finish \tab Finish \tab endyear,endmonth,endday \tab  \tab  \cr
   StateName \tab Name of State \tab State.Name \tab statenme \tab Name \tab  \cr
   cowNR \tab Cow NR. \tab COW.Nr \tab ccode \tab  \tab  \cr
   RatProcedure \tab  \tab  \tab  \tab  \tab Rat \cr
}
}

\keyword{datasets}
