% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-states.R
\docType{data}
\name{states}
\alias{states}
\title{States datacube}
\format{
\describe{
\item{HUGGO: }{A dataset with 470
observations and 14 variables:
stateID, Begin, End, StateName, Capital, Latitude, Longitude, Constitution, RatProcedure, Area, Region, StateName2, Capital2, and Source_rat.}
\item{GW: }{A dataset with 216
observations and 6 variables:
stateID, Begin, End, StateName, cowID, and cowNR.}
\item{ISD: }{A dataset with 362
observations and 8 variables:
stateID, Begin, End, StateName, cowID, cowNR, Micro, and NewState.}
}
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
}
\source{
\describe{
\item{HUGGO: }{
Hollway, James, Henrique Sposito, and Jael Tan. 2021. \emph{manystates: States for manydata}.
}
\item{GW: }{
Gleditsch, K.S., and M. D. Ward. “Interstate system membership: A revised list of the independent states since 1816”.
\emph{International Interactions} 25.4 (1999), pp. 393-413.
\url{http://ksgleditsch.com/data-4.html}
}
\item{ISD: }{
Griffiths, R.D., and C. R. Butcher. “Introducing the international system (s) dataset (ISD), 1816-2011”.
\emph{International Interactions} 39.5 (2013), pp. 748-768.
\url{http://www.ryan-griffiths.com/data}
}
}
}
\usage{
states
}
\description{
The \code{states} datacube is a list containing
3 datasets:
HUGGO, GW, and ISD.
It is a work-in-progress, so please let us know if you have any comments or suggestions.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $HUGGO
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |stateID     |character|    470|          0|         0|
#> |Begin       |mdate    |    320|        150|     31.91|
#> |End         |mdate    |    320|        150|     31.91|
#> |StateName   |character|    468|          2|      0.43|
#> |Capital     |character|    467|          3|      0.64|
#> |Latitude    |character|    467|          3|      0.64|
#> |Longitude   |character|    467|          3|      0.64|
#> |Constitution|character|     30|        440|     93.62|
#> |RatProcedure|character|     30|        440|     93.62|
#> |Area        |character|    211|        259|     55.11|
#> |Region      |character|    211|        259|     55.11|
#> |StateName2  |character|      6|        464|     98.72|
#> |Capital2    |character|      1|        469|     99.79|
#> |Source_rat  |character|     38|        432|     91.91|
#> -------------------------------------------------------
#> 
#> 
#> $GW
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |stateID     |character|    216|          0|         0|
#> |Begin       |mdate    |    216|          0|         0|
#> |End         |mdate    |    216|          0|         0|
#> |StateName   |character|    216|          0|         0|
#> |cowID       |character|    216|          0|         0|
#> |cowNR       |character|    216|          0|         0|
#> -------------------------------------------------------
#> 
#> 
#> $ISD
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |stateID     |character|    362|          0|         0|
#> |Begin       |mdate    |    362|          0|         0|
#> |End         |mdate    |    362|          0|         0|
#> |StateName   |character|    362|          0|         0|
#> |cowID       |character|    362|          0|         0|
#> |cowNR       |character|    362|          0|         0|
#> |Micro       |character|    362|          0|         0|
#> |NewState    |character|    362|          0|         0|
#> -------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{Mapping}{
\tabular{lll}{
   \emph{manystates} \tab \emph{GW} \tab \emph{ISD} \cr
   stateID \tab  \tab  \cr
   Begin \tab Start \tab Start \cr
   End \tab Finish \tab Finish \cr
   StateName \tab Name of State \tab State.Name \cr
   cowID \tab Cow ID \tab COW.ID \cr
   cowNR \tab Cow NR. \tab COW.Nr \cr
}
}

\keyword{datasets}
