% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-states.R
\docType{data}
\name{states}
\alias{states}
\title{States datacube}
\format{
\describe{
\item{ISD: }{A dataset with 499
observations and 36 variables:
stateID, StateName, Begin, End, StateNameAlt, Latitude, Longitude, StartType, EndType, cowID, cowNR, ISD_Category, Region, Start_Am, EStart_Am, Declare, DecDate, Population, ..., VioEnd, and VioEnd_Am.}
\item{HUGGO: }{A dataset with 406
observations and 19 variables:
stateID, StateName, Capital, Begin, Basis, DecIndep, Autonomy, End, Grounds, Latitude, Longitude, Region, StateNameAlt, CapitalAlt, Constitution, RatProcedure, Coder, Comments, and Source.}
\item{GW: }{A dataset with 216
observations and 7 variables:
stateID, StateName, Begin, End, StateNameAlt, cowID, and cowNR.}
}
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
}
\source{
\itemize{
\item Griffiths, Ryan D., and Charles R. Butcher. 'Introducing the international system(s) dataset (ISD), 1816-2011'. International Interactions 39.5 (2013), pp. 748-768.
\item Hollway, James, Henrique Sposito, and Jael Tan. 2021. States for manydata.
\item Gleditsch, Kristian S., and Michael D. Ward. 'Interstate system membership: A revised list of the independent states since 1816'. International Interactions 25.4 (1999), pp. 393-413.
}
}
\usage{
states
}
\description{
The \code{manystates::states} datacube is a list containing 3 datasets: ISD, HUGGO, and GW.
It is a work-in-progress, so please let us know if you have any comments or suggestions.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $ISD
#> ---------------------------------------------------------
#> |   Variable   |  Class  |  Obs  |  Missing  |  Miss \%  |
#> ---------------------------------------------------------
#> |stateID       |character|    499|          0|         0|
#> |StateName     |character|    499|          0|         0|
#> |Begin         |mdate    |    282|        217|     43.49|
#> |End           |mdate    |    499|          0|         0|
#> |StateNameAlt  |character|    210|        289|     57.92|
#> |Latitude      |character|    343|        156|     31.26|
#> |Longitude     |character|    343|        156|     31.26|
#> |StartType     |numeric  |    337|        162|     32.46|
#> |EndType       |numeric  |    296|        203|     40.68|
#> |cowID         |character|    499|          0|         0|
#> |cowNR         |numeric  |    499|          0|         0|
#> |ISD_Category  |numeric  |    497|          2|       0.4|
#> |Region        |numeric  |    499|          0|         0|
#> |Start_Am      |numeric  |    499|          0|         0|
#> |EStart_Am     |numeric  |    232|        267|     53.51|
#> |Declare       |numeric  |    315|        184|     36.87|
#> |DecDate       |character|     72|        427|     85.57|
#> |Population    |character|    319|        180|     36.07|
#> |PopDate       |numeric  |    316|        183|     36.67|
#> |PopAm         |numeric  |    339|        160|     32.06|
#> |PopulationHigh|numeric  |    139|        360|     72.14|
#> |PopulationLow |numeric  |    108|        391|     78.36|
#> |StartType_Am  |numeric  |    339|        160|     32.06|
#> |StartSettle   |numeric  |    320|        179|     35.87|
#> |End_Am        |numeric  |    499|          0|         0|
#> |EndType_Am    |numeric  |    294|        205|     41.08|
#> |EndSettle     |numeric  |    282|        217|     43.49|
#> |Sovereignty_Am|numeric  |    499|          0|         0|
#> |EuroDip       |numeric  |    331|        168|     33.67|
#> |Borders       |numeric  |    332|        167|     33.47|
#> |Borders_Am    |numeric  |    342|        157|     31.46|
#> |Capital       |character|    284|        215|     43.09|
#> |VioStart      |numeric  |    318|        181|     36.27|
#> |VioStart_Am   |numeric  |    327|        172|     34.47|
#> |VioEnd        |numeric  |    292|        207|     41.48|
#> |VioEnd_Am     |numeric  |    297|        202|     40.48|
#> ---------------------------------------------------------
#> 
#> 
#> $HUGGO
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |stateID     |character|    406|          0|         0|
#> |StateName   |character|    406|          0|         0|
#> |Capital     |character|    406|          0|         0|
#> |Begin       |mdate    |    405|          1|      0.25|
#> |Basis       |character|    406|          0|         0|
#> |DecIndep    |mdate    |    406|          0|         0|
#> |Autonomy    |mdate    |    406|          0|         0|
#> |End         |mdate    |    406|          0|         0|
#> |Grounds     |character|    406|          0|         0|
#> |Latitude    |numeric  |    406|          0|         0|
#> |Longitude   |numeric  |    406|          0|         0|
#> |Region      |character|    406|          0|         0|
#> |StateNameAlt|character|     55|        351|     86.45|
#> |CapitalAlt  |character|      9|        397|     97.78|
#> |Constitution|character|     34|        372|     91.63|
#> |RatProcedure|character|     36|        370|     91.13|
#> |Coder       |character|    406|          0|         0|
#> |Comments    |character|    406|          0|         0|
#> |Source      |character|     61|        345|     84.98|
#> -------------------------------------------------------
#> 
#> 
#> $GW
#> -------------------------------------------------------
#> |  Variable  |  Class  |  Obs  |  Missing  |  Miss \%  |
#> -------------------------------------------------------
#> |stateID     |character|    216|          0|         0|
#> |StateName   |character|    216|          0|         0|
#> |Begin       |mdate    |    216|          0|         0|
#> |End         |mdate    |    216|          0|         0|
#> |StateNameAlt|character|     20|        196|     90.74|
#> |cowID       |character|    216|          0|         0|
#> |cowNR       |character|    216|          0|         0|
#> -------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{Mapping}{
\tabular{lll}{
   \emph{manystates} \tab \emph{GW} \tab \emph{ISD} \cr
   stateID \tab  \tab  \cr
   Begin \tab Start \tab Start \cr
   End \tab Finish \tab Finish \cr
   StateName \tab Name of State \tab State.Name \cr
   cowID \tab Cow ID \tab COW.ID \cr
   cowNR \tab Cow NR. \tab COW.Nr \cr
}
}

\keyword{datasets}
