% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manystates-states.R
\docType{data}
\name{states}
\alias{states}
\title{states database documentation}
\format{
The states database is a list that contains the
following 4 datasets: GW, ISD, COW, ICOW.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{GW: }{A dataset with 216 observations and the following
5 variables: GW_ID, Beg, End, COW_Nr, Label.}
\item{ISD: }{A dataset with 362 observations and the following
7 variables: ISD_ID, Beg, End, COW_Nr, Label, Micro, New.State.}
\item{COW: }{A dataset with 243 observations and the following
5 variables: COW_ID, Beg, End, COW_Nr, Label.}
\item{ICOW: }{A dataset with 217 observations and the following
15 variables: ICOW_ID, Label, ColRuler, IndFrom, IndDate, IndViol, IndType, SecFrom, SecDate, SecViol, Into, IntoDate, COWsys, GWsys, Notes.}
}
}
\usage{
states
}
\description{
states database documentation
}
\details{
#``` {r, echo = FALSE, warning = FALSE}
#lapply(states, skimr::skim_without_charts)\preformatted{}
}
\keyword{datasets}
