% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qData-states.R
\docType{data}
\name{states}
\alias{states}
\title{states database documentation}
\format{
The states database is a list that contains the following  datasets: GW, ISD, COW. For more info, please use the report_data() function.
\describe{
\item{GW: A dataset with 216 observations and the following 5 variables: COW_Nr, ID, Beg, End, Label.}{Describe dataset here}
\item{ISD: A dataset with 362 observations and the following 7 variables: COW_Nr, ID, Beg, End, Label, Micro, New.State.}{Describe dataset here}
\item{COW: A dataset with 243 observations and the following 5 variables: COW_Nr, ID, Beg, End, Label.}{Describe dataset here}
}
}
\source{
\url{https://correlatesofwar.org/data-sets/state-system-membership}
}
\usage{
states
}
\description{
Write the database citation here.
}
\keyword{datasets}
